% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{add_lines_X_SIPw}
\alias{add_lines_X_SIPw}
\title{Add lines for the density of infected individuals for a SIPw model}
\usage{
add_lines_X_SIPw(
  XH,
  pars,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X.SIPw}

\item{pars}{a list that defines an `ramp.xde` model (*e.g.*,  generated by `xde_setup()`)}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for a SIPw model
}
