% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{add_lines_X_SIPw_dts}
\alias{add_lines_X_SIPw_dts}
\title{Add lines for the density of infected individuals for the SIPw_dts model}
\usage{
add_lines_X_SIPw_dts(
  XH,
  pars,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X.SIPw_dts}

\item{pars}{a list that defines an `ramp.dts` model (*e.g.*,  generated by `dts_setup()`)}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the SIPw_dts model
}
