% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{BQ}
\alias{BQ}
\title{The \code{BQ} module for the MY component}
\description{
Implements the \strong{MY} component using a BQ (Blood-feeding / egg-laying Queue)
model of adult mosquito ecology and infection dynamics. Mosquitoes alternate
between a blood-feeding state (\strong{B}) and a gravid / egg-laying state (\strong{Q}),
and infection is tracked across both states.
}
\section{State Variables}{

\describe{
\item{\code{Bu}}{density of uninfected, blood-feeding mosquitoes}
\item{\code{Qu}}{density of uninfected, gravid (egg-laying) mosquitoes}
\item{\code{By}}{density of infected (exposed), blood-feeding mosquitoes}
\item{\code{Qy}}{density of infected (exposed), gravid mosquitoes}
\item{\code{Bz}}{density of infectious, blood-feeding mosquitoes}
\item{\code{Qz}}{density of infectious, gravid mosquitoes}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{q}}{human blood fraction}
\item{\code{g}}{mosquito mortality rate}
\item{\code{sigma_b}}{emigration rate during blood feeding}
\item{\code{sigma_q}}{emigration rate during egg laying}
\item{\code{mu}}{fraction lost during emigration}
\item{\code{nu}}{oviposition rate (per mosquito)}
\item{\code{eip}}{extrinsic incubation period (\eqn{\tau}); \eqn{\phi = 1/\tau}}
\item{\code{eggsPerBatch}}{eggs laid per oviposition bout}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dB_u/dt &= \Lambda + \nu Q_u - f B_u - \Omega_b \cdot B_u \\
dQ_u/dt &= f(1 - q\kappa) B_u - \nu Q_u - \Omega_q \cdot Q_u \\
dB_y/dt &= \nu Q_y - (f + \phi) B_y - \Omega_b \cdot B_y \\
dQ_y/dt &= fq\kappa B_u + f B_y - (\nu + \phi) Q_y - \Omega_q \cdot Q_y \\
dB_z/dt &= \phi B_y + \nu Q_z - f B_z - \Omega_b \cdot B_z \\
dQ_z/dt &= \phi Q_y + f B_z - \nu Q_z - \Omega_q \cdot Q_z \\
\end{array}
}
}

