% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_dispersal.R
\name{change_K_matrix.Ks}
\alias{change_K_matrix.Ks}
\title{Change mosquito dispersal matrix \code{Ks}}
\usage{
\method{change_K_matrix}{Ks}(K_matrix, xds_obj, which_K = "Ks", s = 1)
}
\arguments{
\item{K_matrix}{a mosquito dispersal \link{matrix}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{which_K}{which K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change the dispersal
matrix for sugar feeding, called the \code{Ks_matrix}
}
\seealso{
\link[ramp.xds:xds_info_mosquito_dispersal]{ramp.xds::xds_info_mosquito_dispersal}; \link[ramp.xds:setup_K_matrix]{ramp.xds::setup_K_matrix}
}
\keyword{internal}
