% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_dispersal.R
\name{change_K_matrix.q}
\alias{change_K_matrix.q}
\title{Change Mosquito Dispersal Matrix}
\usage{
\method{change_K_matrix}{q}(K_matrix, xds_obj, behave = "q", s = 1)
}
\arguments{
\item{K_matrix}{a mosquito dispersal \link{matrix}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{behave}{the behavioral state}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change the dispersal matrix for
egg laying
}
\seealso{
\link[ramp.xds:xds_info_mosquito_dispersal]{ramp.xds::xds_info_mosquito_dispersal}; \link[ramp.xds:setup_K_matrix]{ramp.xds::setup_K_matrix}
}
\keyword{internal}
