% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{change_L_inits.stages}
\alias{change_L_inits.stages}
\title{Change initial values for \code{stages} (\strong{L})}
\usage{
\method{change_L_inits}{stages}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Initial values of \eqn{E, L_1, L_2, L_3, L_4, P} are reset
if they are passed as named components of \code{options}.
}
\keyword{internal}
