% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BRQS.R
\name{change_MY_inits.BRQS}
\alias{change_MY_inits.BRQS}
\title{Make inits for BRQS adult mosquito model}
\usage{
\method{change_MY_inits}{BRQS}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
none
}
\description{
Make inits for BRQS adult mosquito model
}
\keyword{internal}
