% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_demography.R
\name{change_Omega_q}
\alias{change_Omega_q}
\title{Change Omega}
\usage{
change_Omega_q(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Update the demographic matrix for
behavioral state models
}
\keyword{internal}
