% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{check_L.stages}
\alias{check_L.stages}
\title{Check the \code{stages} module}
\usage{
\method{check_L}{stages}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run no consistency checks
}
\keyword{internal}
