% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{check_XH.SIPd}
\alias{check_XH.SIPd}
\title{Run checks before solving (\strong{XH})}
\usage{
\method{check_XH}{SIPd}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run checks before solving (\strong{XH})
}
\keyword{internal}
