% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{create_Xinits_SEIRV}
\alias{create_Xinits_SEIRV}
\title{Make initial values for the SEIRV human model, with defaults}
\usage{
create_Xinits_SEIRV(
  nStrata,
  Xopts = list(),
  H0 = NULL,
  S = NULL,
  I = 1,
  E = 0,
  R = 1,
  V = 1
)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a [list] to overwrite defaults}

\item{H0}{the initial value for H}

\item{S}{the initial value for S}

\item{I}{the initial value for I}

\item{E}{the initial value for E}

\item{R}{the initial values for R}

\item{V}{the initial values for V}
}
\value{
a [list]
}
\description{
Make initial values for the SEIRV human model, with defaults
}
