% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{create_Xinits_SIPw}
\alias{create_Xinits_SIPw}
\title{Make initial values for a SIPw human model, with defaults}
\usage{
create_Xinits_SIPw(nStrata, H, Xopts = list(), I = 1, P = 0, w = 0)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{H}{the initial human population density}

\item{Xopts}{a [list] that could overwrite defaults}

\item{I}{the initial values of the variable I}

\item{P}{the initial values of the variable P}

\item{w}{the initial values of the tracking variable w}
}
\value{
a [list]
}
\description{
Make initial values for a SIPw human model, with defaults
}
