% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{create_Xinits_SIR}
\alias{create_Xinits_SIR}
\title{Make initial values for the SIR human model, with defaults}
\usage{
create_Xinits_SIR(nStrata, Xopts = list(), H0 = NULL, S = NULL, I = 1, R = 1)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a [list] to overwrite defaults}

\item{H0}{the initial value for H}

\item{S}{the initial value for S}

\item{I}{the initial value for I}

\item{R}{the initial values for R}
}
\value{
a [list]
}
\description{
Make initial values for the SIR human model, with defaults
}
