% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{create_Xpar_SEIRV}
\alias{create_Xpar_SEIRV}
\title{Make parameters for SEIRV human model, with defaults}
\usage{
create_Xpar_SEIRV(
  nStrata,
  Xopts = list(),
  alpha = 0.1,
  b = 0.55,
  r = 1/180,
  c = 0.15,
  tau = 0.5,
  varepsilon = 0.5,
  gamma = 0.5
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{alpha}{proportion of vaccinated humans}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the rate infections clear}

\item{c}{the proportion of bites on infected humans that infect a mosquito}

\item{tau}{incubation rate}

\item{varepsilon}{proportion of recovered humans that are protected from the pathogen}

\item{gamma}{loss of immunity rate}
}
\value{
a [list]
}
\description{
Make parameters for SEIRV human model, with defaults
}
