% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{create_Xpar_SIR}
\alias{create_Xpar_SIR}
\title{Make parameters for SIR human model, with defaults}
\usage{
create_Xpar_SIR(nStrata, Xopts = list(), b = 0.55, r = 1/180, c = 0.15)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the the duration of an infection}

\item{c}{the proportion of bites on infected humans that infect a mosquito}
}
\value{
a [list]
}
\description{
Make parameters for SIR human model, with defaults
}
