% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{create_Xpar_SIRS}
\alias{create_Xpar_SIRS}
\title{Make parameters for SIRS human model, with defaults}
\usage{
create_Xpar_SIRS(
  nStrata,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  c = 0.15,
  gamma = 0.5
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the the duration of an infection}

\item{c}{the proportion of bites on infected humans that infect a mosquito}

\item{gamma}{the rate of loss of immunity}
}
\value{
a [list]
}
\description{
Make parameters for SIRS human model, with defaults
}
