% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{create_Xpar_workhorse}
\alias{create_Xpar_workhorse}
\title{Make parameters for workhorse human model, with defaults}
\usage{
create_Xpar_workhorse(
  nStrata,
  Xopts = list(),
  tau = 12,
  b = 0.55,
  r = 1/200,
  sigma = 1/30,
  phi = 1/10,
  c1 = 0.3,
  c2 = 0.1,
  c3 = 0.03,
  c4 = 0.01,
  cG = 0.15,
  xi_1 = 1/20,
  xi_2 = 1/40,
  xi_3 = 1/80,
  zeta_U = 1/365,
  zeta_A = 1/100,
  zeta_1 = 1/100,
  zeta_2 = 1/180,
  zeta_3 = 1/180,
  zeta_4 = 1/365
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{tau}{the incubation period}

\item{b}{the probability of infection, per infectious bite}

\item{r}{the clearance rate}

\item{sigma}{the rate individuals lose chemoprotection}

\item{phi}{the rate of transition from the acute to the chronic phase}

\item{c1}{the probability of infecting a mosquito, stage 1}

\item{c2}{the probability of infecting a mosquito, stage 2}

\item{c3}{the probability of infecting a mosquito, stage 3}

\item{c4}{the probability of infecting a mosquito, stage 4}

\item{cG}{the probability of infecting a mosquito, prophylaxed}

\item{xi_1}{the rate of transition from stage 1 to stage 2}

\item{xi_2}{the rate of transition from stage 2 to stage 3}

\item{xi_3}{the rate of transition from stage 3 to stage 4}

\item{zeta_U}{the treatment rate for uninfected individuals}

\item{zeta_A}{the treatment rate for acutely infected individuals}

\item{zeta_1}{the treatment rate for stage 1}

\item{zeta_2}{the treatment rate for stage 2}

\item{zeta_3}{the treatment rate for stage 3}

\item{zeta_4}{the treatment rate for stage 4}
}
\value{
a [list]
}
\description{
Make parameters for workhorse human model, with defaults
}
