% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQm.R
\name{dMYdt.BQm}
\alias{dMYdt.BQm}
\title{Compute derivatives for \code{BQm} (\strong{MY})}
\usage{
\method{dMYdt}{BQm}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:dMYdt]{dMYdt} for the BQm ODE model.
}
\keyword{internal}
