% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{dMYdt.SEI}
\alias{dMYdt.SEI}
\title{Compute derivatives for \code{SEI} (\strong{MY})}
\usage{
\method{dMYdt}{SEI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:dMYdt]{dMYdt} for the SEI ODE model.

The dynamics of adult mosquitoes:
\deqn{dM/dt = \Lambda - \Omega \cdot M}

The density of infected but not infectious mosquitoes:
\deqn{dY/dt = fq\kappa(M-Y) - \Omega \cdot Y}

The density of infectious mosquitoes:
\deqn{dZ/dt = (Y-Z)/\tau - \Omega \cdot Z}
}
\keyword{internal}
