% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-macdonald.R
\name{dMYZdt.macdonald}
\alias{dMYZdt.macdonald}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYZdt}{macdonald}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dMYZdt] for Macdonald's DDE model.
}
