% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{dMYZdt.RMG_xde}
\alias{dMYZdt.RMG_xde}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYZdt}{RMG_xde}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dMYZdt] for the RMG_xde ODE model.
}
