% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{dts_setup_MYZpar.ENBRQ_dts}
\alias{dts_setup_MYZpar.ENBRQ_dts}
\title{Setup MYZpar for the ENBRQ_dts model}
\usage{
\method{dts_setup_MYZpar}{ENBRQ_dts}(MYZname, pars, s, EIPopts, MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a [list] vector
}
\description{
Implements [dts_setup_MYZpar] for the ENBRQ_dts model
}
