% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{dts_setup_Xpar.SIPw}
\alias{dts_setup_Xpar.SIPw}
\title{Setup the Xpar for the SIPw_dts model}
\usage{
\method{dts_setup_Xpar}{SIPw}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [dts_setup_Xpar] for a SIPw dts model
}
