% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{dXdt.garki}
\alias{dXdt.garki}
\title{Derivatives for human population}
\usage{
\method{dXdt}{garki}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dXdt] for the garki model.
}
