% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{dXdt.garki_xde}
\alias{dXdt.garki_xde}
\title{Derivatives for human population}
\usage{
\method{dXdt}{garki_xde}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dXdt] for the garki_xde model.
}
