% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{dXdt.SIP}
\alias{dXdt.SIP}
\title{Derivatives for the `SIP` Module (**X** Component)}
\usage{
\method{dXdt}{SIP}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Compute the derivatives for SIP compartmental model:
\deqn{
\begin{array}{rrrrcc}
dS/dt =& - (h +\xi) S &+ r I & + \eta P & + d{\cal H}(S) &+ B(t, H) \\
dI/dt =& (1-\rho) h S & - (r+\xi) I &&+ d{\cal H}(I) \\
dP/dt =& (\rho h +\xi)S & +\xi I & - \eta P &+ d{\cal H}(P)
\end{array}
}
where \eqn{H = S+I+P}; \eqn{B(t, H)} is the
time-dependent birth rate; and the \eqn{d{\cal H}}
operator computes derivatives for the demographic model \eqn{\cal H}.
}
\seealso{
The parameters are defined in [make_Xpar_SIP]
}
