% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{dXdt.SIPmav3}
\alias{dXdt.SIPmav3}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SIPmav3}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SIPmav3-xde model.
}
