% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{dXdt.SIPw_dts}
\alias{dXdt.SIPw_dts}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SIPw_dts}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dXdt] for the SIPw_dts model.
}
