% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIR.R, R/Xmod-SIRnew.R
\name{dXdt.SIRnew}
\alias{dXdt.SIRnew}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SIRnew}(t, y, pars, i)

\method{dXdt}{SIRnew}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector

a [numeric] vector
}
\description{
Implements [dXdt] for the SIR model with demography.

Implements [dXdt] for the SIRnew model
}
