% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{dXdt.SIRS}
\alias{dXdt.SIRS}
\title{Compute the derivatives for parasite infection dynamics in human population strata}
\usage{
\method{dXdt}{SIRS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dXdt] for the SIRS model
}
