% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{dXdt.workhorse}
\alias{dXdt.workhorse}
\title{Derivatives for human population}
\usage{
\method{dXdt}{workhorse}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [dXdt] for the SIS model, no demography.
}
