% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{dXHdt.garki}
\alias{dXHdt.garki}
\title{Derivatives for human population}
\usage{
\method{dXHdt}{garki}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements a continuous time version of the Garki model
}
\references{
{This implements a version of the model
developed for the Garki Project
\insertRef{DietzK1974GarkiModel}{ramp.library}}
}
