% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{dXHdt.SEIR}
\alias{dXHdt.SEIR}
\title{Compute the derivatives for parasite infection dynamics in human population strata}
\usage{
\method{dXHdt}{SEIR}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:dXHdt]{dXHdt} for the SEIR model
}
\keyword{internal}
