% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{dXHdt.SEISd}
\alias{dXHdt.SEISd}
\title{Compute derivatives for the \strong{XH} module \code{SEISd}}
\usage{
\method{dXHdt}{SEISd}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:dXHdt]{dXHdt} for the SEISd model
}
\keyword{internal}
