% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{dXHdt.SIP}
\alias{dXHdt.SIP}
\title{Derivatives for the \code{SIP} Module (\strong{XH} Component)}
\usage{
\method{dXHdt}{SIP}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for SIP compartmental model:
\deqn{
\begin{array}{rl}
dH/dt =& B(t,H) + D \cdot H \\
dI/dt =& (1-\rho) h S  - (r+\xi+\sigma ) I + D \cdot I \\
dP/dt =& \rho h S + \xi (S+I)  + \sigma I - \eta P + D \cdot P
\end{array}
}
where \eqn{S = H-I-P}; \eqn{B(t, H)} is the
time-dependent birth rate; and \eqn{D} the demographic matrix.

The model sets up ports for mass treatment:
\itemize{
\item the parameter \code{xi_t} is computed as \eqn{\xi'(t) = \xi + \mbox{mda}(t)}
\item the parameter \code{sigma_t} is computed as \eqn{\sigma'(t) = \sigma + \mbox{msat}(t)}
}

The functions \code{mda} and \code{msat} are set to zero. To configure mass treatment, see \code{ramp.control}
}
\seealso{
The parameters are defined in \link{make_XH_obj_SIP}
}
