% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{dXHdt.SIPd}
\alias{dXHdt.SIPd}
\title{Compute derivatives for the \strong{XH} module \code{SIPd}}
\usage{
\method{dXHdt}{SIPd}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for SIPd compartmental model:
\deqn{
\begin{array}{rrrrcc}
dH/dt =& B(t,H) + D \cdot H \\
dI/dt =& (1-\rho) h S & - (r+\xi) I &&+ D \cdot I \\
dP/dt =& (\rho h +\xi)S & +\xi I & - \eta P &+ d{\cal H}(P)
\end{array}
}
where \eqn{S = H-I-P}; \eqn{B(t, H)} is the
time-dependent birth rate; and the demographic matrix is \eqn{D}
}
\seealso{
The parameters are defined in \link{make_XH_obj_SIPd}
}
