% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-weekly.R
\name{Exposure.weekly}
\alias{Exposure.weekly}
\title{Exposure and Infection}
\usage{
\method{Exposure}{weekly}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object as a \link{list}}
}
\value{
the function modifies **pars** and returns it: the computed FoI are stored as `pars$FoI`
}
\description{
This function translates seven days of daily entomological
inoculation rate (dEIR) into a weekly attack rate. The daily FoI
is the sum of two terms: 1) a function [F_foi] computes the local dFoI;
2) a function [travel_malaria] that computes the FoI resulting from
exposure while traveling.
}
