% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRV.R
\name{F_b.SEIRV}
\alias{F_b.SEIRV}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_b}{SEIRV}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_b] for the SEIRV model.
}
