% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_clearance.R
\name{F_clear.r}
\alias{F_clear.r}
\title{Natural Parasite Clearance}
\usage{
\method{F_clear}{r}(y, Xpars)
}
\arguments{
\item{y}{state vector}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_clear} for a constant
rate of clearance for simple infections
}
