% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{F_detect.dim}
\alias{F_detect.dim}
\title{Detection}
\usage{
\method{F_detect}{dim}(y, Xpars, mod_detect)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_detect}{an \code{F_detect} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
The basic model for the
risk of transmission, per bloodmeal:
a constant probability of
transmission, per bloodmeal
}
