% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{F_detect_vars.mav}
\alias{F_detect_vars.mav}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_detect_vars}{mav}(vars, Xpars, mod_detect)
}
\arguments{
\item{vars}{variables as a named list}

\item{Xpars}{an \strong{X} model object}

\item{mod_detect}{an detection model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements a model for detection motivated by
parasite densities and detection
}
