% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{F_detect_y.mav}
\alias{F_detect_y.mav}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_detect_y}{mav}(y, Xpars, mod_detect)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_detect}{an \code{F_detect} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements a model for detection motivated
by parasite densities
}
