% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{F_detect_y}
\alias{F_detect_y}
\title{Set up a model for detection}
\usage{
F_detect_y(y, Xpars, mod_detect)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_detect}{an \code{F_detect} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A model could have models
for detection by different methods, so detection
models must be dispatched by passing their
model objects.
}
