% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQm.R
\name{F_eggs.BQm}
\alias{F_eggs.BQm}
\title{Number of eggs laid by adult mosquitoes}
\usage{
\method{F_eggs}{BQm}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link[ramp.xds:F_eggs]{F_eggs} for the BQm model.
}
\keyword{internal}
