% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-SBRQ.R
\name{F_eggs.SBRQ}
\alias{F_eggs.SBRQ}
\title{Number of eggs laid by adult mosquitoes}
\usage{
\method{F_eggs}{SBRQ}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_eggs} for the SBRQ model.
}
