% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{F_emerge.stages}
\alias{F_emerge.stages}
\title{Compute emergent adults for \code{stages} (\strong{L} component)}
\usage{
\method{F_emerge}{stages}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
The number of female adults emerging from the habitats,
per day, is:
\deqn{\psi_P P/2 }
}
\seealso{
\link{dLdt.stages}
}
\keyword{internal}
