% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{F_emerge.stages_dts}
\alias{F_emerge.stages_dts}
\title{Number of newly emerging adults from each larval habitat}
\usage{
\method{F_emerge}{stages_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a [numeric] vector of length `nHabitats`
}
\description{
Implements [F_emerge] for the stages_dts competition model.
}
