% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-dddn.R
\name{F_f.dddn}
\alias{F_f.dddn}
\title{Dawn, day, dusk, night model for the blood feeding rate}
\usage{
\method{F_f}{dddn}(t, vars, f_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{f_par}{a \link{list}}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Implements [F_f] for a dddn model
}
