% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-forcing.R
\name{F_force_malaria.func}
\alias{F_force_malaria.func}
\title{Mass Treatment}
\usage{
\method{F_force_malaria}{func}(t, y, Xpars, forcing_mod)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{forcing_mod}{a function}
}
\value{
a \link{numeric} vector of length ``
}
\description{
Implements forcing for the trivial model.
}
