% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-SBRQ.R
\name{F_fqM.SBRQ}
\alias{F_fqM.SBRQ}
\title{Blood feeding rate of the infective mosquito population}
\usage{
\method{F_fqM}{SBRQ}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Implements [F_fqM] for the SBRQ model.
}
