% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{F_fqZ.ENBRQ_dts}
\alias{F_fqZ.ENBRQ_dts}
\title{The net blood feeding rate of the infective mosquito population in a patch}
\usage{
\method{F_fqZ}{ENBRQ_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Implements [F_fqZ] for the ENBRQ_dts model.
}
