% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{F_fqZ.RMG}
\alias{F_fqZ.RMG}
\title{Blood feeding rate of the infective mosquito population}
\usage{
\method{F_fqZ}{RMG}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link[ramp.xds:F_fqZ]{ramp.xds::F_fqZ} for the RMG model.
}
