% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{F_fqZ.SEI}
\alias{F_fqZ.SEI}
\title{The net blood feeding rate of the infective mosquito population in a patch}
\usage{
\method{F_fqZ}{SEI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link[ramp.xds:F_fqZ]{F_fqZ} for the SEI model.
}
\keyword{internal}
