% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{F_H.SIP}
\alias{F_H.SIP}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SIP}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_I]{F_I} for the SIP model.
}
\keyword{internal}
