% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEIS.R
\name{F_I.SEIS}
\alias{F_I.SEIS}
\title{Size of effective infectious human population}
\usage{
\method{F_I}{SEIS}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_I]{F_I} for the SEIS model.
}
\keyword{internal}
