% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{F_incidence_h.p}
\alias{F_incidence_h.p}
\title{Malaria from Exposure}
\usage{
\method{F_incidence_h}{p}(y, Xpars, mod_ince_h)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_ince_h}{an \code{F_incidence_h} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Computes the probability of
getting malaria from a new infection
}
