% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{F_incidence_h}
\alias{F_incidence_h}
\title{Malaria Incidence from Exposure}
\usage{
F_incidence_h(y, Xpars, mod_ince_h)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_ince_h}{an \code{F_incidence_h} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the probability of
getting malaria as a fraction of new
infections.
}
