% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{F_incidence_x.protect}
\alias{F_incidence_x.protect}
\title{Malaria Incidents, per infection}
\usage{
\method{F_incidence_x}{protect}(y, Xpars, mod_ince_h)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A model for incidence
assuming a \emph{bad types} model
}
