% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{F_incidence_x.rate}
\alias{F_incidence_x.rate}
\title{Basic model for}
\usage{
\method{F_incidence_x}{rate}(y, Xpars, mod_ince_x)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_ince_x}{an \code{F_incidence_x} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements a basic model for
incidence from infection
}
