% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{F_incidence_x}
\alias{F_incidence_x}
\title{Malaria Incidence from Infection}
\usage{
F_incidence_x(y, Xpars, mod_ince_x)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_ince_x}{an \code{F_incidence_x} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the rate that malaria
cases arise from infected states
}
