% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_infect.R
\name{F_infect_im}
\alias{F_infect_im}
\title{Infections per Infectious Bite}
\usage{
F_infect_im(vh, b0, Nb)
}
\arguments{
\item{vh}{cumulative exposure}

\item{b0}{infection, per infectious bite, no immunity}

\item{Nb}{effective number of infections}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
The basic model for the
risk of infection, per infectious bite:
a constant probability of
infection, per infectious bite
}
