% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_infect.R
\name{F_infect_vars}
\alias{F_infect_vars}
\title{Infections per Infectious Bite}
\usage{
F_infect_vars(vars, Xpars)
}
\arguments{
\item{vars}{variables as a named list}

\item{Xpars}{an \code{Xmod} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the probability of
infection, per infectious bite. Implements
models of naturally-acquired or vaccine-induced
pre-erythrocytic immunity
}
