% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{F_infectivity.SEIRV}
\alias{F_infectivity.SEIRV}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_infectivity}{SEIRV}(y, xds_obj, i)
}
\arguments{
\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_infectivity]{F_infectivity} for the SEIRV model.
}
\keyword{internal}
