% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEIS.R
\name{F_infectivity.SEIS}
\alias{F_infectivity.SEIS}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_infectivity}{SEIS}(y, xds_obj, i)
}
\arguments{
\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_infectivity]{F_infectivity} for the SEIS model.
}
\keyword{internal}
