% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_mass_treat.R
\name{F_mass_treat.func}
\alias{F_mass_treat.func}
\title{Mass Treatment}
\usage{
\method{F_mass_treat}{func}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{i}{the species index}
}
\value{
a \link{numeric} vector of length ``
}
\description{
Implements \link{F_mass_treat} for the trivial model.
}
