% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_mass_treat.R
\name{F_mass_treat}
\alias{F_mass_treat}
\title{Mass Treatment}
\usage{
F_mass_treat(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{i}{the species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$mass_treat}.
}
