% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{F_ni.garki}
\alias{F_ni.garki}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_ni}{garki}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_ni]{F_ni} for the SIR model.
}
\keyword{internal}
