% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{F_ni.SEISd}
\alias{F_ni.SEISd}
\title{Compute the NI}
\usage{
\method{F_ni}{SEISd}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_ni]{F_ni} for the SEISd model.
}
\keyword{internal}
