% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{F_ni.SIP}
\alias{F_ni.SIP}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_ni}{SIP}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_ni]{F_ni} for the SIP model.
}
\keyword{internal}
