% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-dddn.R
\name{F_nu.dddn}
\alias{F_nu.dddn}
\title{Dawn, day, dusk, night model for the human fraction}
\usage{
\method{F_nu}{dddn}(t, vars, nu_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{nu_par}{a \link{list}}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Implements [F_nu] for a dddn model
}
