% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-dddn.R
\name{F_p.dddn}
\alias{F_p.dddn}
\title{Dawn, day, dusk, night model for the human fraction}
\usage{
\method{F_p}{dddn}(t, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{MYZpar}{a \link{list}}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Implements [F_p] for a dddn model
}
