% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{F_pfpr_by_rdt.SIPmav3}
\alias{F_pfpr_by_rdt.SIPmav3}
\title{Compute PR by RDT}
\usage{
\method{F_pfpr_by_rdt}{SIPmav3}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for SIPmav3 models
}
