% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{F_pr.SEIR}
\alias{F_pr.SEIR}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SEIR(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEIR model.
}
