% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{F_pr.SEIRV}
\alias{F_pr.SEIRV}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SEIRV(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEIRV model.
}
