% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{F_pr.SEIS}
\alias{F_pr.SEIS}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SEIS(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEIS model.
}
