% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_pr.SEISd}
\alias{F_pr.SEISd}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SEISd(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEISd model.
}
