% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_pr.SIP}
\alias{F_pr.SIP}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SIP(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIP model.
}
