% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{F_pr.SIPd}
\alias{F_pr.SIPd}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SIPd(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIPd model.
}
