% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{F_pr.SIPw}
\alias{F_pr.SIPw}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{SIPw}(varslist, pars, i)
}
\arguments{
\item{varslist}{a list with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_pr] for SIPw models
}
