% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{F_pr.SIR}
\alias{F_pr.SIR}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SIR(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIR model.
}
