% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-workhorse.R
\name{F_pr.workhorse}
\alias{F_pr.workhorse}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{workhorse}(varslist, pars, i)
}
\arguments{
\item{varslist}{a list with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_pr] for the workhorse model.
}
