% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{F_pr_by_pcr.SEIS}
\alias{F_pr_by_pcr.SEIS}
\title{Compute the prevalence of infection by pcr}
\usage{
F_pr_by_pcr.SEIS(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEIS model.
}
