% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_pr_by_pcr.SEISd}
\alias{F_pr_by_pcr.SEISd}
\title{Compute the prevalence of infection by pcr}
\usage{
\method{F_pr_by_pcr}{SEISd}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_pr] for the SEISd model.
}
