% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_pr_by_pcr.SIP}
\alias{F_pr_by_pcr.SIP}
\title{Compute the prevalence of infection by pcr}
\usage{
F_pr_by_pcr.SIP(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIP model.
}
