% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_pr_by_rdt.SEISd}
\alias{F_pr_by_rdt.SEISd}
\title{Compute the prevalence of infection by RDT}
\usage{
F_pr_by_rdt.SEISd(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEISd model.
}
