% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{F_prevalence.garki}
\alias{F_prevalence.garki}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
<<<<<<< HEAD:man/F_pr.garki.Rd
F_pr.garki(vars, Xpar)
=======
\method{F_prevalence}{garki}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
>>>>>>> acc05432180b090912479adcca191d41600b0345:man/F_prevalence.garki.Rd
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the true pr for the garki model.
}
