% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{F_prevalence.SIPd}
\alias{F_prevalence.SIPd}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_prevalence}{SIPd}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_prevalence]{F_prevalence} for the SIPd model.
}
\keyword{internal}
