% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{F_sigma_b}
\alias{F_sigma_b}
\title{Compute the Mosquito Patch Emigration Rate}
\usage{
F_sigma_b(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector osigma length \code{nPatches}
}
\description{
This method dispatches on the type of \code{sigma_obj}. It should
set the values the patch emigration rate, \eqn{\sigma}
}
\keyword{internal}
