% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_drug_taking.R
\name{F_take_drugs.func}
\alias{F_take_drugs.func}
\title{Mass Treatment}
\usage{
\method{F_take_drugs}{func}(t, y, Xpars, take_drugs_mod)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{take_drugs_mod}{a function}
}
\value{
a \link{numeric} vector of length ``
}
\description{
Implements \link{F_take_drugs} for the trivial model.
}
