% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_drug_taking.R
\name{F_take_drugs}
\alias{F_take_drugs}
\title{Mass Treatment}
\usage{
F_take_drugs(t, y, Xpars, take_drugs_mod)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{take_drugs_mod}{a function}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$mass_treat}.
}
