% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_transmit.R
\name{F_transmit_dim}
\alias{F_transmit_dim}
\title{A model for detection}
\usage{
F_transmit_dim(vh, moi, aoi, c0, pda, Nc)
}
\arguments{
\item{vh}{cumulative exposure}

\item{moi}{mean multiplicity of infection}

\item{aoi}{mean age of infection}

\item{c0}{probability of transmition per transmitious cite, no immunity}

\item{pda}{a model for detection by aoi, motivated by a model for parasite densities and detection}

\item{Nc}{a model for protection by exposure}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
The basic model for the
risk of transmission, per bloodmeal:
a constant probability of
transmission, per bloodmeal
}
