% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_transmit.R
\name{F_transmit_vars.c}
\alias{F_transmit_vars.c}
\title{Infectiousness, basic}
\usage{
\method{F_transmit_vars}{c}(vars, Xpars)
}
\arguments{
\item{vars}{variables as a named list}

\item{Xpars}{an \code{Xmod} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Return a constant value, \eqn{c},
for the probability a mosquito would become infected
after blood feeding on a human, by strata
}
