% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_transmit.R
\name{F_transmit_vars}
\alias{F_transmit_vars}
\title{Compute Infectiousness with Named Variables}
\usage{
F_transmit_vars(vars, Xpars)
}
\arguments{
\item{vars}{variables as a named list}

\item{Xpars}{an \code{Xmod} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method computes infectiousness,
the probability a mosquito would become infected
after blood feeding on a human, by strata
}
