% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_transmit.R
\name{F_transmit_y.dim}
\alias{F_transmit_y.dim}
\title{A model for detection}
\usage{
\method{F_transmit_y}{dim}(y, Xpars)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
The basic model for the
risk of transmission, per bloodmeal:
a constant probability of
transmission, per bloodmeal
}
