% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-treatment.R
\name{F_treat.p3}
\alias{F_treat.p3}
\title{Basic Treat and Cure}
\usage{
\method{F_treat}{p3}(y, Xpars, mod_treat)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_treat}{an \code{F_treat} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{treat} for the SIP model.
}
