% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-treatment.R
\name{F_treat}
\alias{F_treat}
\title{Treat and Cure}
\usage{
F_treat(y, Xpars, mod_treat)
}
\arguments{
\item{y}{state vector}

\item{Xpars}{an \code{Xmod} object}

\item{mod_treat}{an \code{F_treat} model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Define a model for the
fraction of infections arising that
are treated and cured. The model should
align with a model for malaria incidence.
}
