% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{F_X.SIPw_dts}
\alias{F_X.SIPw_dts}
\title{Size of effective infectious human population}
\usage{
\method{F_X}{SIPw_dts}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_X] for the SIPw_dts model.
}
