% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{F_X.SIRS}
\alias{F_X.SIRS}
\title{Size of effective infectious human population}
\usage{
\method{F_X}{SIRS}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_X] for the SIS model.
}
