% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{get_f.RMG}
\alias{get_f.RMG}
\title{Get the feeding rate}
\usage{
\method{get_f}{RMG}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Get the feeding rate
}
