% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-SBRQ.R
\name{get_f.SBRQ}
\alias{get_f.SBRQ}
\title{Get the feeding rate}
\usage{
\method{get_f}{SBRQ}(pars, s = 1)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{s}{the vector species index}
}
\value{
a [numeric] vector
}
\description{
Get the feeding rate
}
