% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{get_HTC.SEIR}
\alias{get_HTC.SEIR}
\title{Compute the HTC for the SEIR model}
\usage{
\method{get_HTC}{SEIR}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{get_HTC} for the SEIR model with demography.
}
\keyword{internal}
