% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEIS.R
\name{get_HTC.SEIS}
\alias{get_HTC.SEIS}
\title{Compute the HTC for the SEIS model}
\usage{
\method{get_HTC}{SEIS}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{get_HTC} for the SEIS model with demography.
}
\keyword{internal}
