% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{get_HTC.SEISd}
\alias{get_HTC.SEISd}
\title{Compute the HTC for the SEISd model}
\usage{
\method{get_HTC}{SEISd}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{get_HTC} for the SEISd model with demography.
}
\keyword{internal}
