% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{get_inits_MYZ.ENBRQ_dts}
\alias{get_inits_MYZ.ENBRQ_dts}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{ENBRQ_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
[numeric]
}
\description{
Implements [get_inits_MYZ] for the ENBRQ_dts model.
}
