% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{get_inits_MYZ.RMG_xde}
\alias{get_inits_MYZ.RMG_xde}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{RMG_xde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements [get_inits_MYZ] for the RMG_xde model.
}
