% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{get_inits_X.garki_xde}
\alias{get_inits_X.garki_xde}
\title{Return initial values as a vector for the garki_xde model}
\usage{
\method{get_inits_X}{garki_xde}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a named [list]
}
\description{
This method dispatches on the type of `pars$Xpar`.
}
