% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{get_inits_X.SEIRnew}
\alias{get_inits_X.SEIRnew}
\title{Return initial values as a vector}
\usage{
\method{get_inits_X}{SEIRnew}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
This method dispatches on the type of `pars$Xpar`.
}
