% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{get_inits_X.SIPw_dts}
\alias{get_inits_X.SIPw_dts}
\title{Return initial values as a vector}
\usage{
\method{get_inits_X}{SIPw_dts}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of `pars$Xpar`.
}
