% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_dispersal.R
\name{get_K_matrix.b}
\alias{get_K_matrix.b}
\title{Get the Mosquito Dispersal Matrix}
\usage{
\method{get_K_matrix}{b}(xds_obj, behave = "b", s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{behave}{the behavioral state}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get the Mosquito Dispersal Matrix
}
\keyword{internal}
