% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_dispersal.R
\name{get_K_matrix.Kq}
\alias{get_K_matrix.Kq}
\title{Get the Mosquito Dispersal Matrix}
\usage{
\method{get_K_matrix}{Kq}(xds_obj, which_K = "Kq", s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{which_K}{which K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get the Mosquito Dispersal Matrix
}
\keyword{internal}
