% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{get_L_pars.stages}
\alias{get_L_pars.stages}
\title{Get parameters for \code{stages} (\strong{L})}
\usage{
\method{get_L_pars}{stages}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Get the \strong{L} component parameters
}
\seealso{
\link{dLdt.stages} or \link{change_L_pars.stages}
}
\keyword{internal}
