% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{get_L_vars.stages}
\alias{get_L_vars.stages}
\title{List variables for \code{stages} (\strong{L})}
\usage{
\method{get_L_vars}{stages}(y, xds_obj, s)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a named \link{list}
}
\description{
Extract the \strong{L} component variables from the
vector of state variables (\code{y}) and return them as a named list.
}
\keyword{internal}
