% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{get_Linits.stages_dts}
\alias{get_Linits.stages_dts}
\title{Return initial values as a vector}
\usage{
\method{get_Linits}{stages_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements [get_Linits] for the GeRM model.
}
