% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{get_MYZinits.RMG}
\alias{get_MYZinits.RMG}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{RMG}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements [get_MYZinits] for the RMG model.
}
