% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-SBRQ.R
\name{get_MYZinits.SBRQ}
\alias{get_MYZinits.SBRQ}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{SBRQ}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_MYZinits} for the SBRQ model.
}
