% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{get_MYZpars.SEI}
\alias{get_MYZpars.SEI}
\title{Return the parameters as a list}
\usage{
\method{get_MYZpars}{SEI}(pars, s = 1)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{s}{the vector species index}
}
\value{
a [list]
}
\description{
This method dispatches on the type of `pars$MYZpar[[s]]`.
}
