% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{get_q.RMG}
\alias{get_q.RMG}
\title{Get the feeding rate}
\usage{
\method{get_q}{RMG}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the feeding rate
}
