% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{get_XH_vars.SEIR}
\alias{get_XH_vars.SEIR}
\title{Return the variables as a list}
\usage{
\method{get_XH_vars}{SEIR}(y, xds_obj, i)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj}
}
