% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-garki.R
\name{get_Xinits.garki}
\alias{get_Xinits.garki}
\title{Return initial values as a vector for the garki model}
\usage{
\method{get_Xinits}{garki}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a named [list]
}
\description{
This method dispatches on the type of `pars$Xpar`.
}
