% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{get_Xinits.SEIR}
\alias{get_Xinits.SEIR}
\title{Return initial values as a vector}
\usage{
\method{get_Xinits}{SEIR}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
