% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{get_Xinits.SIPmav3}
\alias{get_Xinits.SIPmav3}
\title{Return initial values as a vector from a SIPmav3 model}
\usage{
\method{get_Xinits}{SIPmav3}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}
}
