% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{get_Xinits.workhorse}
\alias{get_Xinits.workhorse}
\title{Return initial values as a vector}
\usage{
\method{get_Xinits}{workhorse}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
This method dispatches on the type of `pars$Xpar`.
}
