% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{get_Xpars.SIP}
\alias{get_Xpars.SIP}
\title{Return the parameters as a list}
\usage{
\method{get_Xpars}{SIP}(pars, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [list]
}
\description{
Parameter values for the \eqn{i^{th}} host are
stored as `pars$Xpar[[i]]`. This returns the stored parameter
values as a list.
}
\seealso{
[create_Xpar_SIP]
}
