% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{HTC.garki}
\alias{HTC.garki}
\title{Compute the HTC for the garki model}
\usage{
\method{HTC}{garki}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the garki model
}
