% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{HTC.SEIR}
\alias{HTC.SEIR}
\title{Compute the HTC for the SEIR model}
\usage{
\method{HTC}{SEIR}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SEIR model with demography.
}
