% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{HTC.SEIRnew}
\alias{HTC.SEIRnew}
\title{Compute the HTC for the SEIRnew model}
\usage{
\method{HTC}{SEIRnew}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SEIRnew model with demography.
}
