% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{HTC.SEIRV}
\alias{HTC.SEIRV}
\title{Compute the HTC for the SEIRV model}
\usage{
\method{HTC}{SEIRV}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SEIRV model with demography.
}
