% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{HTC.SEIS}
\alias{HTC.SEIS}
\title{Compute the HTC for the SEIS model}
\usage{
\method{HTC}{SEIS}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SEIS model with demography.
}
