% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{HTC.SIP}
\alias{HTC.SIP}
\title{Compute the HTC for the SIP model}
\usage{
\method{HTC}{SIP}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIP model with demography.
}
