% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{HTC.SIPmav3}
\alias{HTC.SIPmav3}
\title{Compute the HTC for the SIPmav3_xde model}
\usage{
\method{HTC}{SIPmav3}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIPmav3_xde model with demography.
}
