% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-xde-SIPw.R
\name{HTC.SIPw}
\alias{HTC.SIPw}
\title{Compute the HTC for the SIPw_xde model}
\usage{
\method{HTC}{SIPw}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIPw_xde model with demography.
}
