% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{HTC.SIPw_dts}
\alias{HTC.SIPw_dts}
\title{Compute the HTC for the SIPw_dts model}
\usage{
\method{HTC}{SIPw_dts}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SIPw_dts model with demography.
}
