% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{HTC.SIRnew}
\alias{HTC.SIRnew}
\title{Compute the HTC for the SIRnew model}
\usage{
\method{HTC}{SIRnew}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SIRnew model with demography.
}
