% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRS.R
\name{HTC.SIRS}
\alias{HTC.SIRS}
\title{Compute the HTC for the SIRS model}
\usage{
\method{HTC}{SIRS}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SIRS model with demography.
}
