% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{HTC.workhorse}
\alias{HTC.workhorse}
\title{Compute the HTC for the workhorse model}
\usage{
\method{HTC}{workhorse}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the workhorse model with demography.
}
