% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{LEffectSizes.stages}
\alias{LEffectSizes.stages}
\title{Apply effect sizes for \code{stages} (\strong{L})}
\usage{
\method{LEffectSizes}{stages}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link[ramp.xds:LEffectSizes]{LEffectSizes} for the \code{stages} model.
Computes effective parameter values as the product of each time-varying
baseline (\verb{_t}) and its corresponding effect size (\code{es_}).
}
\keyword{internal}
