% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{list_Lvars.stages_dts}
\alias{list_Lvars.stages_dts}
\title{Return the variables as a list}
\usage{
\method{list_Lvars}{stages_dts}(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a [list]
}
\description{
This method dispatches on the type of `pars$Lpar[[s]]`
}
