% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{list_Xvars.SEIRnew}
\alias{list_Xvars.SEIRnew}
\title{Return the variables as a list}
\usage{
\method{list_Xvars}{SEIRnew}(y, pars, i)
}
\arguments{
\item{y}{the variables}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [list]
}
\description{
This method dispatches on the type of `pars$Xpar`
}
