% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{list_Xvars.SIR}
\alias{list_Xvars.SIR}
\title{Return the variables as a list}
\usage{
\method{list_Xvars}{SIR}(y, pars, i)
}
\arguments{
\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar}
}
