% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{make_indices_MYZ.RMG_xde}
\alias{make_indices_MYZ.RMG_xde}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{RMG_xde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements [make_indices_MYZ] for the RMG_xde model.
}
