% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{make_indices_X.garki_xde}
\alias{make_indices_X.garki_xde}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{garki_xde}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [make_indices_X] for the garki_xde model.
}
