% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{make_indices_X.SEIRnew}
\alias{make_indices_X.SEIRnew}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{SEIRnew}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [make_indices_X] for the SEIRnew model.
}
