% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{make_indices_X.SIPw_dts}
\alias{make_indices_X.SIPw_dts}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{SIPw_dts}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [make_indices_X] for the SIPw_dts model.
}
