% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-dts-stages.R
\name{make_inits_L_stages_dts}
\alias{make_inits_L_stages_dts}
\title{Make inits for stages_dts competition aquatic mosquito model}
\usage{
make_inits_L_stages_dts(pars, L1, L2, L3, L4, P)
}
\arguments{
\item{pars}{a \link{list}}

\item{L1}{initial conditions}

\item{L2}{initial conditions}

\item{L3}{initial conditions}

\item{L4}{initial conditions}

\item{P}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Make inits for stages_dts competition aquatic mosquito model
}
