% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-dts-ENBRQ.R
\name{make_inits_MYZ_ENBRQ_dts}
\alias{make_inits_MYZ_ENBRQ_dts}
\title{Make inits for ENBRQ_dts adult mosquito model}
\usage{
make_inits_MYZ_ENBRQ_dts(pars, E, N, B, Q, R1, R2)
}
\arguments{
\item{pars}{a \link{list}}

\item{E}{total mosquito density at each patch}

\item{N}{total mosquito density at each patch}

\item{B}{blood feeding mosquito density at each patch}

\item{Q}{egg laying mosquito density at each patch}

\item{R1}{total stage 1 resting mosquitoes in each patch}

\item{R2}{total stage 2 resting mosquitoes in each patch}
}
\value{
a \link{list}
}
\description{
Make inits for ENBRQ_dts adult mosquito model
}
