% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{make_L_inits_stages}
\alias{make_L_inits_stages}
\title{Make initial values for \code{stages} (\strong{L} component)}
\usage{
make_L_inits_stages(
  nHabitats,
  options = list(),
  E = 1,
  L1 = 1,
  L2 = 1,
  L3 = 1,
  L4 = 1,
  P = 1
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{options}{a \link{list} that overwrites default values}

\item{E}{initial egg density}

\item{L1}{initial first instar density}

\item{L2}{initial second instar density}

\item{L3}{initial third instar density}

\item{L4}{initial fourth instar density}

\item{P}{initial pupal density}
}
\value{
a \link{list} with initial conditions
}
\description{
Initial values of \eqn{E, L_1, L_2, L_3, L_4, P} can be
set by passing the corresponding names in \code{options}.
}
\keyword{internal}
