% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{make_L_obj_stages}
\alias{make_L_obj_stages}
\title{Make \code{L_obj} for \code{stages} (\strong{L} component)}
\usage{
make_L_obj_stages(
  nHabitats,
  options = list(),
  psi_E = 1/2,
  phi_E = 1/10,
  psi1 = 1/2,
  xi1 = 0,
  phi1 = 1/2,
  theta1 = 1/100,
  psi2 = 1/2,
  xi2 = 0,
  phi2 = 1/2,
  theta2 = 1/100,
  psi3 = 1/2,
  xi3 = 0,
  phi3 = 1/2,
  theta3 = 1/100,
  psi4 = 1/2,
  xi4 = 0,
  phi4 = 1/2,
  theta4 = 1/100,
  phi_P = 1/2,
  psi_P = 1/3,
  w1 = 0.3,
  w2 = 0.5,
  w3 = 0.8
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{options}{a named \link{list}}

\item{psi_E}{egg maturation (hatching) rate}

\item{phi_E}{egg mortality rate}

\item{psi1}{first instar maturation rate}

\item{xi1}{first instar delayed maturation response to crowding}

\item{phi1}{first instar density-independent mortality rate}

\item{theta1}{first instar density-dependent mortality slope}

\item{psi2}{second instar maturation rate}

\item{xi2}{second instar delayed maturation response to crowding}

\item{phi2}{second instar density-independent mortality rate}

\item{theta2}{second instar density-dependent mortality slope}

\item{psi3}{third instar maturation rate}

\item{xi3}{third instar delayed maturation response to crowding}

\item{phi3}{third instar density-independent mortality rate}

\item{theta3}{third instar density-dependent mortality slope}

\item{psi4}{fourth instar maturation rate}

\item{xi4}{fourth instar delayed maturation response to crowding}

\item{phi4}{fourth instar density-independent mortality rate}

\item{theta4}{fourth instar density-dependent mortality slope}

\item{phi_P}{pupal mortality rate}

\item{psi_P}{pupal maturation (emergence) rate}

\item{w1}{contribution of \eqn{L_1} to mean crowding}

\item{w2}{contribution of \eqn{L_2} to mean crowding}

\item{w3}{contribution of \eqn{L_3} to mean crowding}
}
\value{
\strong{\code{L_obj}} an \strong{L} component object
}
\description{
The following parameters will be set to the values in
\code{options}. If they are not found, default values will be used.
\itemize{
\item \eqn{\psi_E} or \code{psi_E}: egg maturation (hatching) rate
\item \eqn{\phi_E} or \code{phi_E}: egg mortality rate
\item \eqn{\psi_i} or \code{psi1}, \code{psi2}, \code{psi3}, \code{psi4}: larval instar maturation rates
\item \eqn{\xi_i} or \code{xi1}, \code{xi2}, \code{xi3}, \code{xi4}: delayed maturation responses to crowding
\item \eqn{\phi_i} or \code{phi1}, \code{phi2}, \code{phi3}, \code{phi4}: density-independent larval mortality rates
\item \eqn{\theta_i} or \code{theta1}, \code{theta2}, \code{theta3}, \code{theta4}: density-dependent larval mortality slopes
\item \eqn{\phi_P} or \code{phi_P}: pupal mortality rate
\item \eqn{\psi_P} or \code{psi_P}: pupal maturation (emergence) rate
\item \eqn{w_1, w_2, w_3} or \code{w1}, \code{w2}, \code{w3}: instar crowding weights
}
}
\seealso{
Called by: \link{setup_L_obj.stages}. Related: \link{dLdt.stages}
}
\keyword{internal}
