% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{make_Linits.stages_dts}
\alias{make_Linits.stages_dts}
\title{Setup the stages_dts model}
\usage{
\method{make_Linits}{stages_dts}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{an \code{xds} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a [list]
}
\description{
Implements [make_Linits] for the stages_dts model
}
