% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-dts-stages.R
\name{make_Linits_stages_dts}
\alias{make_Linits_stages_dts}
\title{Make inits for stages_dts competition aquatic mosquito model}
\usage{
make_Linits_stages_dts(
  nHabitats,
  Lopts = list(),
  L1 = 1,
  L2 = 0,
  L3 = 0,
  L4 = 0,
  P = 0
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{L1}{initial conditions}

\item{L2}{initial conditions}

\item{L3}{initial conditions}

\item{L4}{initial conditions}

\item{P}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Make inits for stages_dts competition aquatic mosquito model
}
