% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{make_Lpar.stages_dts}
\alias{make_Lpar.stages_dts}
\title{Setup Lpar for the stages_dts model}
\usage{
\method{make_Lpar}{stages_dts}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a named \link{list} to configure \strong{\code{Lpar}}}
}
\value{
a [list] vector
}
\description{
Implements [make_Lpar] for the stages_dts model
}
