% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_mass_treat.R
\name{make_mass_test_treat}
\alias{make_mass_test_treat}
\title{Make a function to simulate mass treatment}
\usage{
make_mass_test_treat(
  nStrata,
  opts = list(),
  scale = 1,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list()
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{opts}{a \link{list} that overwrites default values}

\item{scale}{a scaling parameter}

\item{F_season}{a function describing a seasonal pattern over time}

\item{season_par}{an object to configure a seasonality function using \link{make_function}}

\item{F_trend}{a function describing a temporal trend over time}

\item{trend_par}{an object to configure a trends function using \link{make_function}}
}
\value{
none
}
\description{
Mass treatment
}
