% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{make_MY_inits_BQ}
\alias{make_MY_inits_BQ}
\title{Make inits for BQ adult mosquito model}
\usage{
make_MY_inits_BQ(
  nPatches,
  options = list(),
  Bu0 = 5,
  Qu0 = 1,
  By0 = 1,
  Qy0 = 1,
  Bz0 = 1,
  Qz0 = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{Bu0}{total uninfected, not gravid mosquito density at each patch}

\item{Qu0}{total uninfected, gravid mosquito density at each patch}

\item{By0}{infected, not gravid mosquito density at each patch}

\item{Qy0}{infected, gravid mosquito density at each patch}

\item{Bz0}{infectious, not gravid mosquito density at each patch}

\item{Qz0}{infectious, gravid mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for BQ adult mosquito model
}
\keyword{internal}
