% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{make_MYZinits.ENBRQ_dts}
\alias{make_MYZinits.ENBRQ_dts}
\title{Setup initial values for the ENBRQ_dts model}
\usage{
\method{make_MYZinits}{ENBRQ_dts}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a [list]
}
\description{
Implements [make_MYZinits] for the ENBRQ_dts model
}
