% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{make_MYZinits.RMG}
\alias{make_MYZinits.RMG}
\title{Setup initial values for the RMG model}
\usage{
\method{make_MYZinits}{RMG}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a [list]
}
\description{
Implements [make_MYZinits] for the RM model
}
