% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{make_MYZinits_ENBRQ_dts}
\alias{make_MYZinits_ENBRQ_dts}
\title{Make inits for ENBRQ_dts adult mosquito model}
\usage{
make_MYZinits_ENBRQ_dts(
  nPatches,
  nR1,
  MYZopts = list(),
  E = 5,
  N = 5,
  B = 5,
  Q = 1,
  R1 = 0,
  R2 = 0
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{nR1}{number of time steps in R1}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{E}{total null-parous, blood feeding mosquito density}

\item{N}{total null-parous, blood feeding mosquito density}

\item{B}{total parous, blood feeding mosquito density at each patch}

\item{Q}{total egg laying mosquito density at each patch}

\item{R1}{resting, stage 1, mosquito density at each patch}

\item{R2}{resting, stage 2, mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for ENBRQ_dts adult mosquito model
}
