% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-SBRQ.R
\name{make_MYZinits_SBRQ}
\alias{make_MYZinits_SBRQ}
\title{Make inits for SBRQ adult mosquito model}
\usage{
make_MYZinits_SBRQ(
  nPatches,
  MYZopts = list(),
  Su = 100,
  Bu = 100,
  Ru = 1,
  Qu = 10,
  Sy = 0,
  By = 0,
  Ry = 0,
  Qy = 0,
  Sz = 0,
  Bz = 0,
  Rz = 0,
  Qz = 0
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{Su}{uninfected, sugar feeding mosquitoes}

\item{Bu}{uninfected, blood feeding mosquitoes}

\item{Ru}{uninfected, resting mosquitoes}

\item{Qu}{uninfected, gravid mosqutioes}

\item{Sy}{infected, sugar feeding mosquitoes}

\item{By}{infected, blood feeding mosquitoes}

\item{Ry}{infected, resting mosquitoes}

\item{Qy}{infected, gravid mosqutioes}

\item{Sz}{infectious, sugar feeding mosquitoes}

\item{Bz}{infectious, blood feeding mosquitoes}

\item{Rz}{infectious, resting mosquitoes}

\item{Qz}{infectious, gravid mosqutioes}
}
\value{
a \link{list}
}
\description{
Make inits for SBRQ adult mosquito model
}
