% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{make_MYZpar.RMG}
\alias{make_MYZpar.RMG}
\title{Setup MYZpar for the RMG model}
\usage{
\method{make_MYZpar}{RMG}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_MYZpar] for the RM model
}
