% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{make_MYZpar_RMG_xde}
\alias{make_MYZpar_RMG_xde}
\title{Make parameters for RM ODE adult mosquito model}
\usage{
make_MYZpar_RMG_xde(
  nPatches,
  MYZopts = list(),
  EIPopts,
  calK,
  g = 1/12,
  sigma_b = 1/8,
  sigma_q = 1/8,
  f = 0.5,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZopts}{a [list] of values that overwrites the defaults}

\item{EIPopts}{a [list] that defines the EIP model}

\item{calK}{a mosquito dispersal matrix of dimensions `nPatches` by `nPatches`}

\item{g}{mosquito mortality rate}

\item{sigma_b}{emigration rate while blood feeding}

\item{sigma_q}{emigration rate while egg laying}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a [list]
}
\description{
Make parameters for RM ODE adult mosquito model
}
