% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{make_parameters_MYZ_RMG}
\alias{make_parameters_MYZ_RMG}
\title{Make parameters for RMG ODE adult mosquito model}
\usage{
make_parameters_MYZ_RMG(
  pars,
  g,
  sigma_b,
  sigma_q,
  f,
  q,
  nu,
  eggsPerBatch,
  eip,
  calK
)
}
\arguments{
\item{pars}{a [list]}

\item{g}{mosquito mortality rate}

\item{sigma_b}{emigration rate}

\item{sigma_q}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{eip}{length of extrinsic incubation period}

\item{calK}{mosquito dispersal matrix of dimensions `nPatches` by `nPatches`}
}
\value{
none
}
\description{
Make parameters for RMG ODE adult mosquito model
}
