% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{make_X_indices.SIR}
\alias{make_X_indices.SIR}
\title{Add indices for human population to parameter list}
\usage{
\method{make_X_indices}{SIR}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [make_X_indices] for the SIR model.
}
