% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{make_X_indices.workhorse}
\alias{make_X_indices.workhorse}
\title{Add indices for human population to parameter list}
\usage{
\method{make_X_indices}{workhorse}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [make_X_indices] for the workhorse model.
}
