% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{make_XH_inits_SEISd}
\alias{make_XH_inits_SEISd}
\title{Make initial values for the SEISd Update_XHt human model, with defaults}
\usage{
make_XH_inits_SEISd(nStrata, H, options = list(), E = 0, I = 1)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial human population density}

\item{options}{a \link{list} to overwrite defaults}

\item{E}{the initial values of the parameter E}

\item{I}{the initial values of the parameter I}
}
\value{
a \link{list}
}
\description{
Make initial values for the SEISd Update_XHt human model, with defaults
}
