% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{make_XH_inits_SIPd}
\alias{make_XH_inits_SIPd}
\title{Make initial values for the SIPd human model, with defaults}
\usage{
make_XH_inits_SIPd(nStrata, H, options = list(), I = 1, P = 0)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{H}{the initial human population density}

\item{options}{a \link{list} that could overwrite defaults}

\item{I}{the initial values of the parameter I}

\item{P}{the initial values of the parameter P}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIPd human model, with defaults
}
