% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{make_XH_inits_SIRS}
\alias{make_XH_inits_SIRS}
\title{Make initial values for the SIRS human model, with defaults}
\usage{
make_XH_inits_SIRS(nStrata, H, options = list(), I = 1, R = 0)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial value for H}

\item{options}{a \link{list} to overwrite defaults}

\item{I}{the initial value for I}

\item{R}{the initial values for R}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIRS human model, with defaults
}
