% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{make_XH_obj_SEISd}
\alias{make_XH_obj_SEISd}
\title{Make parameters for SEISd Update_XHt human model, with defaults}
\usage{
make_XH_obj_SEISd(
  nStrata,
  options = list(),
  b = 0.55,
  r = 1/180,
  nu = 14,
  c = 0.15
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{options}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{nu}{1/latent period}

\item{c}{transmission probability (efficiency) from human to mosquito}
}
\value{
a \link{list}
}
\description{
Make parameters for SEISd Update_XHt human model, with defaults
}
