% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{make_Xinits.garki}
\alias{make_Xinits.garki}
\title{Setup Xinits.garki}
\usage{
\method{make_Xinits}{garki}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xinits] for the garki model
}
