% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{make_Xinits.SIPw}
\alias{make_Xinits.SIPw}
\title{Setup Xinits.SIPw}
\usage{
\method{make_Xinits}{SIPw}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
<<<<<<< HEAD
=======

\item{H}{initial host population density}
>>>>>>> 6424052ff1354a233a6cea897d2fe92a200a89e7

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xinits] for the SIPw models
}
