% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{make_Xinits.workhorse}
\alias{make_Xinits.workhorse}
\title{Setup Xinits.workhorse}
\usage{
\method{make_Xinits}{workhorse}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xinits] for the workhorse model
}
