% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{make_Xinits_SEIR}
\alias{make_Xinits_SEIR}
\title{Make initial values for the SEIR human model, with defaults}
\usage{
make_Xinits_SEIR(nStrata, H, Xopts = list(), I = 1, E = 0, R = 1)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial value for H}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{I}{the initial value for I}

\item{E}{the initial value for E}

\item{R}{the initial values for R}
}
\value{
a \link{list}
}
\description{
Make initial values for the SEIR human model, with defaults
}
