% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRV.R
\name{make_Xinits_SEIRV}
\alias{make_Xinits_SEIRV}
\title{Make initial values for the SEIRV human model, with defaults}
\usage{
make_Xinits_SEIRV(
  nStrata,
  Xopts = list(),
  H0 = NULL,
  S0 = NULL,
  I0 = 1,
  E0 = 0,
  R0 = 1,
  V0 = 1
)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a [list] to overwrite defaults}

\item{H0}{the initial value for H}

\item{S0}{the initial value for S}

\item{I0}{the initial value for I}

\item{E0}{the initial value for E}

\item{R0}{the initial values for R}

\item{V0}{the initial values for V}
}
\value{
a [list]
}
\description{
Make initial values for the SEIRV human model, with defaults
}
