% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{make_Xinits_SIPmav3}
\alias{make_Xinits_SIPmav3}
\title{Make initial values for a SIPmav3 human model, with defaults}
\usage{
make_Xinits_SIPmav3(
  nStrata,
  H,
  Xopts = list(),
  I = 1,
  P = 0,
  moi = 0,
  aoi = 0,
  vh = 0,
  severe = 0,
  moderate = 0,
  mild = 0,
  treat = 0
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{H}{the initial human population density}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{I}{the initial values of the variable I}

\item{P}{the initial values of the variable P}

\item{moi}{the initial values of MoI}

\item{aoi}{initial values of AoI}

\item{vh}{initial values for cumulative exposure}

\item{severe}{initial values for cumulative severe malaria}

\item{moderate}{initial values for cumulative moderate malaria}

\item{mild}{initial values for cumulative moderate malaria}

\item{treat}{initial values for cumulative treatment}
}
\value{
a \link{list}
}
\description{
Make initial values for a SIPmav3 human model, with defaults
}
