% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{make_Xinits_SIRnew}
\alias{make_Xinits_SIRnew}
\title{Make initial values for the SIRnew human model, with defaults}
\usage{
make_Xinits_SIRnew(
  nStrata,
  Xopts = list(),
  H0 = NULL,
  S0 = NULL,
  I0 = 1,
  R0 = 1
)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a [list] to overwrite defaults}

\item{H0}{the initial value for H}

\item{S0}{the initial value for S}

\item{I0}{the initial value for I}

\item{R0}{the initial values for R}
}
\value{
a [list]
}
\description{
Make initial values for the SIRnew human model, with defaults
}
