% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{make_Xinits_workhorse}
\alias{make_Xinits_workhorse}
\title{Make initial values for the workhorse human model, with defaults}
\usage{
make_Xinits_workhorse(
  nStrata,
  Xopts = list(),
  H0 = NULL,
  U = 180,
  A0 = 13,
  P = 11,
  G = 3,
  I1 = 20,
  I2 = 40,
  I3 = 80,
  I4 = 100,
  A1 = 2,
  A2 = 4,
  A3 = 8,
  A4 = 12,
  w = 0
)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a [list] to overwrite defaults}

\item{H0}{the initial population size}

\item{U}{the initial number uninfected}

\item{A0}{the initial number in U and A}

\item{P}{the initial number chemoprotected and not infectious}

\item{G}{the initial number chemoprotected and infectious}

\item{I1}{the initial number with the youngest chronic phase infection in stage 1}

\item{I2}{the initial number with the youngest chronic phase infection in stage 2}

\item{I3}{the initial number with the youngest chronic phase infection in stage 3}

\item{I4}{the initial number with the youngest chronic phase infection in stage $}

\item{A1}{the initial number in I1 and A}

\item{A2}{the initial number in I2 and A}

\item{A3}{the initial number in I3 and A}

\item{A4}{the initial number in I4 and A}

\item{w}{exposure tracking variable}
}
\value{
a [list]
}
\description{
Make initial values for the workhorse human model, with defaults
}
