% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{make_Xpar.garki}
\alias{make_Xpar.garki}
\title{make Xpar.garki}
\usage{
\method{make_Xpar}{garki}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xpar] for the garki model
}
