% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{make_Xpar.SEIR}
\alias{make_Xpar.SEIR}
\title{Setup Xpar.SEIR}
\usage{
\method{make_Xpar}{SEIR}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xpar] for the SEIR model
}
