% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{make_Xpar.SIPw}
\alias{make_Xpar.SIPw}
\title{Setup the Xpar for the SIPw_xde model}
\usage{
\method{make_Xpar}{SIPw}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
implements [make_Xpar] for the SIPw model
}
