% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{make_Xpar.workhorse}
\alias{make_Xpar.workhorse}
\title{make Xpar.workhorse}
\usage{
\method{make_Xpar}{workhorse}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xpar] for the workhorse model
}
