% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{make_Xpar_SEIR}
\alias{make_Xpar_SEIR}
\title{Make parameters for SEIR human model, with defaults}
\usage{
make_Xpar_SEIR(
  nStrata,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  c = 0.15,
  tau = 0.5
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the the duration of an infection}

\item{c}{the proportion of bites on infected humans that infect a mosquito}

\item{tau}{incubation rate}
}
\value{
a \link{list}
}
\description{
Make parameters for SEIR human model, with defaults
}
