% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{make_Xpar_SEIRnew}
\alias{make_Xpar_SEIRnew}
\title{Make parameters for SEIRnew human model, with defaults}
\usage{
make_Xpar_SEIRnew(
  nStrata,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  c = 0.15,
  tau = 0.5
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the the duration of an infection}

\item{c}{the proportion of bites on infected humans that infect a mosquito}

\item{tau}{incubation rate}
}
\value{
a [list]
}
\description{
Make parameters for SEIRnew human model, with defaults
}
