% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRV.R
\name{make_Xpar_SEIRV}
\alias{make_Xpar_SEIRV}
\title{Make parameters for SEIRV human model, with defaults}
\usage{
make_Xpar_SEIRV(
  nStrata,
  Xopts = list(),
  alpha = 0.1,
  b = 0.55,
  r = 1/180,
  c = 0.15,
  tau = 0.5,
  eps,
  sig = 0.3,
  delta = 0.5
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{alpha}{travallers with immunity from malaria}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the the duration of an infection}

\item{c}{the proportion of bites on infected humans that infect a mosquito}

\item{tau}{incubation rate}

\item{eps}{loss of immunity rate}

\item{sig}{progression rate of recovered individual to protected class}

\item{delta}{proportion of protected after recovery}
}
\value{
a [list]
}
\description{
Make parameters for SEIRV human model, with defaults
}
