% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{make_Xpar_SIPd}
\alias{make_Xpar_SIPd}
\title{Make parameters for SIPd human model, with defaults}
\usage{
make_Xpar_SIPd(
  nStrata,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  c = 0.15,
  rho = 0.1,
  eta = 25,
  xi = 1/365
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{rho}{probability of successful treatment upon infection}

\item{eta}{duration of chemo-protection}

\item{xi}{background treatment rate}
}
\value{
a [list]
}
\description{
Make parameters for SIPd human model, with defaults
}
