% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{make_Xpar_SIPmav3}
\alias{make_Xpar_SIPmav3}
\title{Make parameters for SIPmav3_xde human model, with defaults}
\usage{
make_Xpar_SIPmav3(
  nStrata,
  Xopts = list(),
  mod_b = "b",
  b_opts = list(b = 0.55),
  mod_c = "c",
  c_opts = list(c = 0.15),
  mod_r = "r",
  r_opts = list(r = 1/180),
  mod_lm = "d",
  lm_opts = list(d = 0.7),
  mod_pcr = "d",
  pcr_opts = list(d = 0.9),
  mod_rdt = "d",
  rdt_opts = list(d = 0.7),
  mod_severe_h = "p",
  severe_h_opts = list(p = 1/500),
  mod_severe_x = "rate",
  severe_x_opts = list(rate = 1/365/5),
  mod_moderate_h = "p",
  moderate_h_opts = list(p = 1/50),
  mod_moderate_x = "rate",
  moderate_x_opts = list(rate = 1/730),
  mod_mild_h = "p",
  mild_h_opts = list(p = 1/10),
  mod_mild_x = "rate",
  mild_x_opts = list(rate = 1/50),
  mod_treat = "p3",
  treat_opts = list(p_sev = 0.95, p_mod = 0.1, p_mild = 0.01),
  mod_xi = "val",
  xi_opts = list(scale = 0),
  mod_seek_care = "val",
  care_seeking_opts = list(val = 0),
  mod_other_fever = "val",
  other_fever_opts = list(val = 0),
  mod_mda = "val",
  mda_opts = list(val = 0),
  mod_msat = "val",
  msat_opts = list(val = 0),
  eta = 1/25
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{mod_b}{a string, the S3 class name for F_infect}

\item{b_opts}{a list: options for setup_F_infect}

\item{mod_c}{a string, the S3 class name for F_transmit}

\item{c_opts}{a list: options for setup_F_transmit}

\item{mod_r}{a string, the S3 class name for F_clear}

\item{r_opts}{a list: options for setup_F_clear}

\item{mod_lm}{a string, the S3 class name for F_detect for light microscopy}

\item{lm_opts}{a list: options for setup_F_detect}

\item{mod_pcr}{a string, the S3 class name for F_detect for PCR}

\item{pcr_opts}{a list: options for setup_F_detect}

\item{mod_rdt}{a string, the S3 class name for F_detect for RDT}

\item{rdt_opts}{a list: options for setup_F_detect}

\item{mod_severe_h}{a string, the S3 class name for F_incidence_h for severe malaria}

\item{severe_h_opts}{a list: options for setup_F_incidence_h}

\item{mod_severe_x}{a string, the S3 class name for F_incidence_x for severe malaria}

\item{severe_x_opts}{a list: options for setup_F_incidence_x}

\item{mod_moderate_h}{a string, the S3 class name for F_incidence_h for moderate malaria}

\item{moderate_h_opts}{a list: options for setup_F_incidence_h}

\item{mod_moderate_x}{a string, the S3 class name for F_incidence_x for moderate malaria}

\item{moderate_x_opts}{a list: options for setup_F_incidence_x}

\item{mod_mild_h}{a string, the S3 class name for F_incidence_h for mild malaria}

\item{mild_h_opts}{a list: options for setup_F_incidence_h}

\item{mod_mild_x}{a string, the S3 class name for F_incidence_x for mild malaria}

\item{mild_x_opts}{a list: options for setup_F_incidence_x}

\item{mod_treat}{a string, the S3 class name for treating malaria}

\item{treat_opts}{a list: options for treating malaria}

\item{mod_xi}{a string, the S3 class name for background drug taking}

\item{xi_opts}{a list: options for setup_F_incidence_x}

\item{mod_seek_care}{a string, the S3 class name for care seeking}

\item{care_seeking_opts}{a list: options to set up care seeking}

\item{mod_other_fever}{a string, the S3 class name for other fever}

\item{other_fever_opts}{a list: options to set up other fever}

\item{mod_mda}{a string, the S3 class name for mass drug administration}

\item{mda_opts}{a list: options to set up mass drug administration}

\item{mod_msat}{a string, the S3 class name for mass screen and treat}

\item{msat_opts}{a list: options to set up mass screen and treat}

\item{eta}{rate of loss of chemo-protection}
}
\value{
a \link{list}
}
\description{
Make parameters for SIPmav3_xde human model, with defaults
}
