% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{make_Xpar_SIPw}
\alias{make_Xpar_SIPw}
\title{Make parameters for SIPw_xde human model, with defaults}
\usage{
make_Xpar_SIPw(
  nStrata,
  Xopts = list(),
  b = 0.55,
  c = 0.15,
  r = 1/180,
  rho = 0.1,
  sigma = 1/730,
  xi = 1/365,
  eta = 1/25
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}

\item{rho}{probability of successful treatment upon infection}

\item{sigma}{probability of treatment for treated, above background}

\item{xi}{background treatment rate}

\item{eta}{rate of loss of chemo-protection}
}
\value{
a [list]
}
\description{
Make parameters for SIPw_xde human model, with defaults
}
