% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{make_Xpar_SIRnew}
\alias{make_Xpar_SIRnew}
\title{Make parameters for SIRnew human model, with defaults}
\usage{
make_Xpar_SIRnew(nStrata, Xopts = list(), b = 0.55, r = 1/180, c = 0.15)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{b}{the proportion of infective bites that cause an infection}

\item{r}{the the duration of an infection}

\item{c}{the proportion of bites on infected humans that infect a mosquito}
}
\value{
a [list]
}
\description{
Make parameters for SIRnew human model, with defaults
}
