% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-SBRQ.R
\name{MBaseline.SBRQ}
\alias{MBaseline.SBRQ}
\title{Reset bloodfeeding and mortality rates to baseline}
\usage{
\method{MBaseline}{SBRQ}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a named [list]
}
\description{
Implements [MBaseline] for the SBRQ model
}
