% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{MBionomics.RMG}
\alias{MBionomics.RMG}
\title{Reset bloodfeeding and mortality rates to baseline}
\usage{
\method{MBionomics}{RMG}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{MBionomics} for the RMG model
}
