% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQm.R
\name{MEffectSizes.BQm}
\alias{MEffectSizes.BQm}
\title{Apply effect sizes for \code{BQm} (\strong{MY})}
\usage{
\method{MEffectSizes}{BQm}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link[ramp.xds:MEffectSizes]{MEffectSizes} for the BQm model
}
\keyword{internal}
