% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BRQS.R
\name{MEffectSizes.BRQS}
\alias{MEffectSizes.BRQS}
\title{Apply effect sizes for \code{BRQS} (\strong{MY})}
\usage{
\method{MEffectSizes}{BRQS}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link[ramp.xds:MEffectSizes]{MEffectSizes} for the BRQS model
}
\keyword{internal}
