% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{parse_L_orbits.stages}
\alias{parse_L_orbits.stages}
\title{Parse outputs for \code{stages} (\strong{L})}
\usage{
\method{parse_L_orbits}{stages}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Returns the columns representing \eqn{E, L_1, L_2, L_3,
L_4, P} from a matrix where each column is a state variable. The
variables are returned as a named list.
}
\keyword{internal}
