% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{parse_Lorbits.stages_dts}
\alias{parse_Lorbits.stages_dts}
\title{Parse the variable names for the stages_dts model}
\usage{
\method{parse_Lorbits}{stages_dts}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
[list]
}
\description{
Implements [parse_Lorbits] for stages_dts competition model.
}
