% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{parse_MY_orbits.RMG}
\alias{parse_MY_orbits.RMG}
\title{parse the output of deSolve and return variables for the RMG model}
\usage{
\method{parse_MY_orbits}{RMG}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{parse_MY_orbits} for the RMG model
}
