% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{parse_MY_orbits.SEI}
\alias{parse_MY_orbits.SEI}
\title{parse the output of deSolve and return variables for the SEI model}
\usage{
\method{parse_MY_orbits}{SEI}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Implements \link[ramp.xds:parse_MY_orbits]{parse_MY_orbits} for the SEI model
}
\keyword{internal}
