% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{parse_MYZorbits.RMG}
\alias{parse_MYZorbits.RMG}
\title{Parse the output of deSolve and return variables for the RMG model}
\usage{
\method{parse_MYZorbits}{RMG}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements [parse_MYZorbits] for the RMG model
}
