% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{parse_outputs_MYZ.RMG}
\alias{parse_outputs_MYZ.RMG}
\title{Parse the output of deSolve and return variables for the RMG model}
\usage{
parse_outputs_MYZ.RMG(outputs, pars, s)
}
\value{
none
}
\description{
Implements [parse_outputs_MYZ] for the RMG model
}
