% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{parse_outputs_X.garki}
\alias{parse_outputs_X.garki}
\title{Parse the output of deSolve and return variables for the garki model}
\usage{
parse_outputs_X.garki(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for the garki model
}
