% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{parse_outputs_X.SEIR}
\alias{parse_outputs_X.SEIR}
\title{Parse the output of deSolve and return variables for the SEIR model}
\usage{
parse_outputs_X.SEIR(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for the SEIR model
}
