% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{parse_outputs_X.SEIRV}
\alias{parse_outputs_X.SEIRV}
\title{Parse the output of deSolve and return variables for the SEIRV model}
\usage{
parse_outputs_X.SEIRV(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for the SEIRV model
}
