% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{parse_outputs_X.SIPw}
\alias{parse_outputs_X.SIPw}
\title{Parse the output of deSolve and return variables for SIPw models}
\usage{
parse_outputs_X.SIPw(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for SIPw models
}
