% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{parse_outputs_X.SIR}
\alias{parse_outputs_X.SIR}
\title{Parse the output of deSolve and return variables for the SIR model}
\usage{
parse_outputs_X.SIR(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for the SIR model
}
