% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{parse_outputs_X.SIRS}
\alias{parse_outputs_X.SIRS}
\title{Parse the output of deSolve and return variables for the SIRS model}
\usage{
parse_outputs_X.SIRS(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for the SIRS model
}
