% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{parse_outputs_X.workhorse}
\alias{parse_outputs_X.workhorse}
\title{Parse the output of deSolve and return variables for the workhorse model}
\usage{
parse_outputs_X.workhorse(outputs, pars, i)
}
\value{
none
}
\description{
Implements [parse_outputs_X] for the workhorse model
}
