% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEIS.R
\name{parse_XH_orbits.SEIS}
\alias{parse_XH_orbits.SEIS}
\title{parse the output of deSolve and return variables for the SEIS model}
\usage{
\method{parse_XH_orbits}{SEIS}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:parse_XH_orbits]{parse_XH_orbits} for the SEIS model
}
\keyword{internal}
