% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{parse_Xorbits.SEIR}
\alias{parse_Xorbits.SEIR}
\title{Parse the output of deSolve and return variables for the SEIR model}
\usage{
<<<<<<< HEAD:man/parse_outputs_X.SEIR.Rd
parse_outputs_X.SEIR(outputs, pars, i)
=======
\method{parse_Xorbits}{SEIR}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
>>>>>>> 6424052ff1354a233a6cea897d2fe92a200a89e7:man/parse_Xorbits.SEIR.Rd
}
\value{
none
}
\description{
Implements [parse_Xorbits] for the SEIR model
}
