% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{parse_Xorbits.SIPmav3}
\alias{parse_Xorbits.SIPmav3}
\title{Parse the output of deSolve and return variables for SIPmav3 models}
\usage{
\method{parse_Xorbits}{SIPmav3}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_Xorbits} for SIPmav3 models
}
