% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{SEI}
\alias{SEI}
\title{The \code{SEI} module for the MY component}
\description{
Implements the \strong{MY} component using a Susceptible-Exposed-Infectious
(SEI) compartmental model of adult mosquito infection dynamics.
}
\section{State Variables}{

\describe{
\item{\code{M}}{density of adult mosquitoes}
\item{\code{Y}}{density of infected (exposed) adult mosquitoes}
\item{\code{Z}}{density of infectious adult mosquitoes}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{q}}{human blood fraction}
\item{\code{g}}{mosquito mortality rate}
\item{\code{sigma}}{patch emigration rate}
\item{\code{mu}}{emigration-related loss}
\item{\code{eip}}{extrinsic incubation period (\eqn{\tau})}
\item{\code{nu}}{oviposition rate (per mosquito)}
\item{\code{eggsPerBatch}}{eggs laid per oviposition bout}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\
dY/dt &= fq\kappa(M-Y) - \Omega \cdot Y \\
dZ/dt &= (Y-Z)/\tau - \Omega \cdot Z \\
\end{array}
}
where \eqn{\Omega} is the demographic matrix encoding mortality and dispersal,
and \eqn{\kappa} is the net infectiousness of humans.
}

