% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{SEIR}
\alias{SEIR}
\title{The \code{SEIR} module for the XH component}
\description{
Implements the \strong{XH} component using a Susceptible-Exposed-Infectious-Recovered
(SEIR) compartmental model of human infection dynamics.
}
\section{State Variables}{

\describe{
\item{\code{H}}{total human (or host) population density}
\item{\code{E}}{density of exposed (infected but not yet infectious) humans}
\item{\code{I}}{density of infectious humans}
\item{\code{R}}{density of recovered humans}
}
Note: susceptible density \eqn{S = H - E - I - R}.
}

\section{Parameters}{

\describe{
\item{\code{b}}{transmission probability from mosquito to human}
\item{\code{c}}{transmission probability from human to mosquito}
\item{\code{tau}}{rate of progression from exposed to infectious}
\item{\code{r}}{clearance rate for infections}
\item{\code{B}}{time-dependent birth rate function \eqn{B(t, H)}}
\item{\code{D}}{linear operator (matrix) for mortality, migration, aging, and transfers}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dH/dt &= B(t,H) + D \cdot H \\
dE/dt &= hS - \tau E + D \cdot E \\
dI/dt &= \tau E - rI + D \cdot I \\
dR/dt &= rI + D \cdot R \\
\end{array}}
where \eqn{h} is the force of infection and \eqn{S = H - E - I - R}.
}

