% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-RMG.R
\name{set_MYZinits.RMG}
\alias{set_MYZinits.RMG}
\title{Set new MYZ parameter values}
\usage{
\method{set_MYZinits}{RMG}(pars, s = 1, MYZopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{MYZopts}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{pars$MYZpar[[s]]}.
}
