% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{set_MYZinits.SEI}
\alias{set_MYZinits.SEI}
\title{Set new MYZ parameter values}
\usage{
\method{set_MYZinits}{SEI}(pars, s = 1, MYZopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{MYZopts}{a named list}
}
\value{
an `xds` object
}
\description{
This method dispatches on the type of `pars$MYZpar[[s]]`.
}
