% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{set_Xinits.SEIS}
\alias{set_Xinits.SEIS}
\title{Return the parameters as a list}
\usage{
\method{set_Xinits}{SEIS}(pars, i = 1, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{Xopts}{a named list}
}
\value{
an **`xds`** object
}
\description{
This method dispatches on the type of `pars$Xpar[[i]]`.
}
