% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{set_Xinits.SIR}
\alias{set_Xinits.SIR}
\title{Return the parameters as a list}
\usage{
\method{set_Xinits}{SIR}(pars, i = 1, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{Xopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
