% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{set_Xpars.SEIR}
\alias{set_Xpars.SEIR}
\title{Return the parameters as a list}
\usage{
\method{set_Xpars}{SEIR}(pars, i = 1, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{Xopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
