% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_clearance.R
\name{setup_F_clear}
\alias{setup_F_clear}
\title{Setup an \code{F_clear} model object}
\usage{
setup_F_clear(mod_clear, r_opts, Xpars, nStrata)
}
\arguments{
\item{mod_clear}{is a string that defines the S3 class}

\item{r_opts}{a set of options to override the defaults}

\item{Xpars}{an \code{Xmod} object}

\item{nStrata}{the number of population strata}
}
\value{
an \code{Xmod} object
}
\description{
Setup an \code{F_clear} model object
}
