% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{setup_F_detect.d}
\alias{setup_F_detect.d}
\title{Set up a basic model for detection}
\usage{
\method{setup_F_detect}{d}(mod_detect, detect_opts, nStrata)
}
\arguments{
\item{mod_detect}{a string, the S3 \code{class} for \code{F_detect}}

\item{detect_opts}{options to set up the model}

\item{nStrata}{the number of population strata}
}
\value{
an \code{F_detect} model object
}
\description{
Set up a \code{F_detect.d} model object,
a constant fraction detected
}
