% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{setup_F_detect.dim}
\alias{setup_F_detect.dim}
\title{Setup an \code{F_detect.dim} model object}
\usage{
\method{setup_F_detect}{dim}(mod_d, d_opts, Xpars, nStrata)
}
\arguments{
\item{Xpars}{an \code{Xmod} object}
}
\value{
an \code{F_detect} model object
}
\description{
Set up a model object
for \link{F_detect.dim}
}
