% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_detect.R
\name{setup_F_detect}
\alias{setup_F_detect}
\title{Set up default detect}
\usage{
setup_F_detect(mod_detect, detect_opts, nStrata)
}
\arguments{
\item{mod_detect}{a string, the S3 \code{class} for \code{F_detect}}

\item{detect_opts}{options to set up the model}

\item{nStrata}{the number of population strata}
}
\value{
an \code{F_detect} model object
}
\description{
Set up an \code{F_detect} model object
}
