% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_infect.R
\name{setup_F_infect.b}
\alias{setup_F_infect.b}
\title{Setup an \code{F_infect.b} model object}
\usage{
\method{setup_F_infect}{b}(mod_b, b_opts, Xpars, nStrata)
}
\arguments{
\item{mod_b}{is a string that defines the S3 class}

\item{b_opts}{a set of options to override the defaults}

\item{Xpars}{an \code{Xmod} object}

\item{nStrata}{the number of population strata}
}
\value{
an \code{F_infect.b} model object
}
\description{
Set up a model object
for the linear model for F_infect
}
