% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_infect.R
\name{setup_F_infect_im}
\alias{setup_F_infect_im}
\title{Setup an \code{F_infect.b} model object}
\usage{
setup_F_infect_im(b0, Nb, nStrata)
}
\arguments{
\item{b0}{probability of infection per infectious bite, no immunity}

\item{Nb}{the number of effective types}

\item{nStrata}{the number of population strata}
}
\value{
an \code{F_infect.b} model object
}
\description{
Set up a F_infect model
with immunity of the form \deqn{b_0 e^{-v_h/N_b}}
}
