% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_transmit.R
\name{setup_F_transmit.c}
\alias{setup_F_transmit.c}
\title{Set up default F_transmit}
\usage{
\method{setup_F_transmit}{c}(mod_c, c_opts, Xpars, nStrata)
}
\arguments{
\item{mod_c}{is a string that defines the S3 class}

\item{c_opts}{a set of options to override the defaults}

\item{Xpars}{an \code{Xmod} object}

\item{nStrata}{the number of population strata}
}
\value{
an \code{Xmod} model object
}
\description{
Implements \link{F_transmit} for the SIP model.
}
