% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_transmit.R
\name{setup_F_transmit_dim}
\alias{setup_F_transmit_dim}
\title{Setup an \code{F_transmit.dim} model object}
\usage{
setup_F_transmit_dim(c0, pda, Nc, nStrata)
}
\arguments{
\item{c0}{probability of transmition per transmitious cite, no immunity}

\item{pda}{a model for detection by aoi, motivated by a model for parasite densities and detection}

\item{Nc}{a model for protection by exposure}

\item{nStrata}{the number of population strata}
}
\value{
an \code{F_transmit.c} model object
}
\description{
Set up a F_transmit model
with immunity of the form \deqn{c_0 e^{-v_h/N_c}}
}
