% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-treatment.R
\name{setup_F_treat.p}
\alias{setup_F_treat.p}
\title{Set up a basic F_treat}
\usage{
\method{setup_F_treat}{p}(mod_F_treat, treat_opts, nStrata)
}
\arguments{
\item{treat_opts}{a set of options to override the defaults}

\item{nStrata}{the number of population strata}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Set up model objects for a basic F_treat model
}
