% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-treatment.R
\name{setup_F_treat}
\alias{setup_F_treat}
\title{Set up F_treat}
\usage{
setup_F_treat(mod_treat, treat_opts, nStrata)
}
\arguments{
\item{mod_treat}{is a string that defines the S3 class}

\item{treat_opts}{a set of options to override the defaults}

\item{nStrata}{the number of population strata}
}
\value{
an \code{Xmod} object
}
\description{
Set up model objects for F_treat
}
