% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-forcing.R
\name{setup_force_malaria.func}
\alias{setup_force_malaria.func}
\title{Make a function to simulate mass treatment}
\usage{
\method{setup_force_malaria}{func}(mod_forcing, nStrata, opts = list())
}
\arguments{
\item{mod_forcing}{is a string that defines an S3 class}

\item{nStrata}{the number of population strata}

\item{opts}{a \link{list} that overwrites default values}
}
\value{
a forcing function
}
\description{
Mass treatment
}
