% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-forcing.R
\name{setup_force_malaria_val}
\alias{setup_force_malaria_val}
\title{Make a function to simulate mass treatment}
\usage{
setup_force_malaria_val(nStrata, opts = list(), value = 0)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{opts}{a \link{list} that overwrites default values}

\item{value}{the value to return}
}
\value{
none
}
\description{
Mass treatment
}
