% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{setup_incidence_h.p}
\alias{setup_incidence_h.p}
\title{Setup Malaria from Exposure, Basic}
\usage{
\method{setup_incidence_h}{p}(mod_ince_h, malaria_opts, nStrata)
}
\arguments{
\item{mod_ince_h}{is a string that defines the S3 class}

\item{malaria_opts}{a set of options to override the defaults}

\item{nStrata}{the number of population strata}
}
\description{
Set up an \code{F_incidence_h.p} model object

an \code{F_incidence_h} model object
}
