% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{setup_incidence_h.protect}
\alias{setup_incidence_h.protect}
\title{Setup an \code{F_incidence_h.protect} model object}
\usage{
\method{setup_incidence_h}{protect}(mod_ince_h, ince_opts, nStrata)
}
\value{
an \code{F_incidence_h.b} model object
}
\description{
Set up a model object
for \link{F_incidence_h.protect}
}
