% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{setup_incidence_x.rate}
\alias{setup_incidence_x.rate}
\title{Setup Malaria from Infection, Basic}
\usage{
\method{setup_incidence_x}{rate}(mod_ince_x, malaria_opts, nStrata)
}
\arguments{
\item{mod_ince_x}{is a string that defines the S3 class}

\item{malaria_opts}{a set of options to override the defaults}

\item{nStrata}{the number of population strata}
}
\description{
Set up a basic model for malaria
incidence from infection, a constant
rate of malaria per infection.

an \code{F_incidence_x} model object
}
