% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{setup_incidence_x}
\alias{setup_incidence_x}
\title{Setup Malaria from Infection}
\usage{
setup_incidence_x(mod_ince_x, malaria_opts, nStrata)
}
\arguments{
\item{mod_ince_x}{is a string that defines the S3 class}

\item{malaria_opts}{a set of options to override the defaults}

\item{nStrata}{the number of population strata}
}
\description{
Set up a \code{F_incidence_x} model object.

an \code{F_incidence_x} model object
}
