% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_incidence.R
\name{setup_incidence_x_protect}
\alias{setup_incidence_x_protect}
\title{Setup an \code{F_incidence_h.protect} model object}
\usage{
setup_incidence_x_protect(rate0, Np, nStrata)
}
\arguments{
\item{rate0}{probability of malaria, per infection}

\item{Np}{the number of effective types}

\item{nStrata}{the number of population strata}
}
\value{
an \code{F_incidence_x.protect} model object
}
\description{
Set up a F_incidence_h model
with immunity of the form \deqn{p_0 e^{-v_h/N_p}}
}
