% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{setup_L_inits.stages}
\alias{setup_L_inits.stages}
\title{Setup initial values for \code{stages} (\strong{L})}
\usage{
\method{setup_L_inits}{stages}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a named \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets initial values of the variables \eqn{E, L_1, L_2,
L_3, L_4, P} by calling \link{make_L_inits_stages}. Default values are used
unless other values are passed in \code{options} by name.
}
\seealso{
\link{make_L_inits_stages}
}
\keyword{internal}
