% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{setup_L_ix.stages}
\alias{setup_L_ix.stages}
\title{Setup variable indices for \code{stages} (\strong{L} component)}
\usage{
\method{setup_L_ix}{stages}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of the indices for the \strong{L} component
variables for the \code{stages} module. Indices are allocated in the order
\eqn{E, L_1, L_2, L_3, L_4, P}, each block of length \code{nHabitats}.
}
\keyword{internal}
