% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{setup_Lix.stages_dts}
\alias{setup_Lix.stages_dts}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{setup_Lix}{stages_dts}(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements [setup_Lix] for stages_dts competition model.
}
