% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{setup_MY_inits.SEI}
\alias{setup_MY_inits.SEI}
\title{Setup initial values for the SEI model}
\usage{
\method{setup_MY_inits}{SEI}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link[ramp.xds:setup_MY_inits]{setup_MY_inits} for the SEI model
}
\keyword{internal}
