% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{setup_MY_ix.BQ}
\alias{setup_MY_ix.BQ}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{BQ}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:setup_MY_ix]{setup_MY_ix} for the BQ model.
}
\keyword{internal}
