% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQm.R
\name{setup_MY_obj.BQm}
\alias{setup_MY_obj.BQm}
\title{Setup MY_obj for the BQm model}
\usage{
\method{setup_MY_obj}{BQm}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link[ramp.xds:setup_MY_obj]{setup_MY_obj} for the RM model
}
\keyword{internal}
