% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{setup_MY_obj.RMG}
\alias{setup_MY_obj.RMG}
\title{Setup MY_obj for the RMG model}
\usage{
\method{setup_MY_obj}{RMG}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_obj} for the RM model
}
