% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{setup_MYZinits.RMG_xde}
\alias{setup_MYZinits.RMG_xde}
\title{Setup initial values for the RMG_xde model}
\usage{
\method{setup_MYZinits}{RMG_xde}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a [list]
}
\description{
Implements [setup_MYZinits] for the RM model
}
