% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-SBRQ.R
\name{setup_MYZinits.SBRQ}
\alias{setup_MYZinits.SBRQ}
\title{Setup initial values for the SBRQ model}
\usage{
\method{setup_MYZinits}{SBRQ}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the RM model
}
