% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-SEI.R
\name{setup_MYZinits.SEI}
\alias{setup_MYZinits.SEI}
\title{Setup initial values for the SEI model}
\usage{
\method{setup_MYZinits}{SEI}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the SEI model
}
