% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{setup_MYZix.ENBRQ_dts}
\alias{setup_MYZix.ENBRQ_dts}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{ENBRQ_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a [list]
}
\description{
Implements [setup_MYZix] for the ENBRQ_dts model.
}
