% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-RMG.R
\name{setup_MYZix.RMG}
\alias{setup_MYZix.RMG}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{RMG}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{setup_MYZix} for the RMG model.
}
