% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-SBRQ.R
\name{setup_MYZix.SBRQ}
\alias{setup_MYZix.SBRQ}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{SBRQ}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{setup_MYZix} for the SBRQ model.
}
