% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{setup_MYZpar.ENBRQ_dts}
\alias{setup_MYZpar.ENBRQ_dts}
\title{Setup MYZpar for the ENBRQ_dts model}
\usage{
\method{setup_MYZpar}{ENBRQ_dts}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the ENBRQ_dts model
}
