% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{setup_MYZpar.RMG_xde}
\alias{setup_MYZpar.RMG_xde}
\title{Setup MYZpar for the RMG_xde model}
\usage{
\method{setup_MYZpar}{RMG_xde}(MYZname, pars, s, EIPopts, MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a [list] vector
}
\description{
Implements [setup_MYZpar] for the RM model
}
