% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-SBRQ.R
\name{setup_MYZpar.SBRQ}
\alias{setup_MYZpar.SBRQ}
\title{Setup MYZpar for the SBRQ model}
\usage{
\method{setup_MYZpar}{SBRQ}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the RM model
}
