% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{setup_MYZpar.SEI}
\alias{setup_MYZpar.SEI}
\title{Setup MYZpar for the SEI model}
\usage{
\method{setup_MYZpar}{SEI}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_MYZpar] for the SEI model
}
