% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{setup_sigma_b_obj}
\alias{setup_sigma_b_obj}
\title{Setup a Patch Emigration Bionomic Object}
\usage{
setup_sigma_b_obj(sigma_b, MY_obj)
}
\arguments{
\item{sigma_b}{the mosquito patch emigration rate}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{sigma}
}
\keyword{internal}
