% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{setup_skillset_L.stages}
\alias{setup_skillset_L.stages}
\title{The \strong{L} module skill set}
\usage{
\method{setup_skillset_L}{stages}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the \strong{\code{xds}} object
}
\description{
The \strong{L} skill set is a list of
a module's capabilities
}
\keyword{internal}
