% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{setup_XH_inits.SEIR}
\alias{setup_XH_inits.SEIR}
\title{Setup XH_obj.SEIR}
\usage{
\method{setup_XH_inits}{SEIR}(xds_obj, H, i, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link[ramp.xds:setup_XH_inits]{setup_XH_inits} for the SEIR model
}
\keyword{internal}
