% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{setup_XH_inits.SIPd}
\alias{setup_XH_inits.SIPd}
\title{Setup Initial Values}
\usage{
\method{setup_XH_inits}{SIPd}(xds_obj, H, i, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Setup Initial Values
}
