% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{setup_XH_ix.SIPd}
\alias{setup_XH_ix.SIPd}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_XH_ix}{SIPd}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:setup_XH_ix]{setup_XH_ix} for the SIPd model.
}
\keyword{internal}
