% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{setup_XH_ix.SIR}
\alias{setup_XH_ix.SIR}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_XH_ix}{SIR}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:setup_XH_ix]{setup_XH_ix} for the SIR model.
}
\keyword{internal}
