% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEIS.R
\name{setup_XH_obj.SEIS}
\alias{setup_XH_obj.SEIS}
\title{Setup XH_obj.SEIS}
\usage{
\method{setup_XH_obj}{SEIS}(Xname, residence, HPop, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{residence}{the residence vector}

\item{HPop}{the initial human population size}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
a \link{list} vector
}
\description{
Implements \link[ramp.xds:setup_XH_obj]{setup_XH_obj} for the SEIS model
}
\keyword{internal}
