% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{setup_Xinits.garki_xde}
\alias{setup_Xinits.garki_xde}
\title{Setup Xinits.garki_xde}
\usage{
\method{setup_Xinits}{garki_xde}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the garki_xde model
}
