% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{setup_Xinits.SEIRnew}
\alias{setup_Xinits.SEIRnew}
\title{Setup Xinits.SEIRnew}
\usage{
\method{setup_Xinits}{SEIRnew}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the SEIRnew model
}
