% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRV.R
\name{setup_Xinits.SEIRV}
\alias{setup_Xinits.SEIRV}
\title{Setup Xinits.SEIRV}
\usage{
\method{setup_Xinits}{SEIRV}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the SEIRV model
}
