% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{setup_Xinits.SIPw_dts}
\alias{setup_Xinits.SIPw_dts}
\title{Setup Xinits.SIPw_dts}
\usage{
\method{setup_Xinits}{SIPw_dts}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the SIPw_dts model
}
