% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{setup_Xinits.SIR}
\alias{setup_Xinits.SIR}
\title{Setup Xinits.SIR}
\usage{
\method{setup_Xinits}{SIR}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xinits} for the SIR model
}
