% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{setup_Xinits.SIRnew}
\alias{setup_Xinits.SIRnew}
\title{Setup Xinits.SIRnew}
\usage{
\method{setup_Xinits}{SIRnew}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the SIRnew model
}
