% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{setup_Xinits.SIRS}
\alias{setup_Xinits.SIRS}
\title{Setup Xinits.SIRS}
\usage{
\method{setup_Xinits}{SIRS}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the SIRS model
}
