% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{setup_Xix.SEISd}
\alias{setup_Xix.SEISd}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{SEISd}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [setup_Xix] for the SEISd model.
}
