% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{setup_Xix.SIP}
\alias{setup_Xix.SIP}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{SIP}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [setup_Xix] for the SIP model.
}
