% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{setup_Xix.SIPmav3}
\alias{setup_Xix.SIPmav3}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{SIPmav3}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{setup_Xix} for SIPmav3 models
}
