% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{setup_Xix.SIPw}
\alias{setup_Xix.SIPw}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{SIPw}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [setup_Xix] for SIPw models
}
