% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{setup_Xix.SIR}
\alias{setup_Xix.SIR}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{SIR}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [setup_Xix] for the SIR model.
}
