% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{setup_Xix.workhorse}
\alias{setup_Xix.workhorse}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{workhorse}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{setup_Xix} for the workhorse model.
}
