% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{setup_Xpar.SEIRV}
\alias{setup_Xpar.SEIRV}
\title{Setup Xpar.SEIRV}
\usage{
\method{setup_Xpar}{SEIRV}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the SEIRV model
}
