% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{setup_Xpar.SEIS}
\alias{setup_Xpar.SEIS}
\title{Setup Xpar.SEIS}
\usage{
\method{setup_Xpar}{SEIS}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xpar] for the SEIS model
}
