% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{setup_Xpar.SIPmav3}
\alias{setup_Xpar.SIPmav3}
\title{Setup the Xpar for the SIPmav3_xde model}
\usage{
\method{setup_Xpar}{SIPmav3}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
implements \link{setup_Xpar} for the SIPmav3 model
}
