% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-xde-SIPw.R
\name{setup_Xpar.SIPw}
\alias{setup_Xpar.SIPw}
\title{Setup the Xpar for the SIPw_xde model}
\usage{
\method{setup_Xpar}{SIPw}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
implements \link{setup_Xpar} for the SIPw model
}
