% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{setup_Xpar.SIPw_dts}
\alias{setup_Xpar.SIPw_dts}
\title{Setup Xpar.SIPw_dts}
\usage{
\method{setup_Xpar}{SIPw_dts}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xpar] for the SIPw_dts model
}
