% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{setup_Xpar.SIRnew}
\alias{setup_Xpar.SIRnew}
\title{Setup Xpar.SIRnew}
\usage{
\method{setup_Xpar}{SIRnew}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xpar] for the SIRnew model
}
