% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-workhorse.R
\name{setup_Xpar.workhorse}
\alias{setup_Xpar.workhorse}
\title{Setup Xpar.workhorse}
\usage{
\method{setup_Xpar}{workhorse}(workhorse, pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xpar] for the workhorse model
}
