% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{SIP}
\alias{SIP}
\title{The \code{SIP} module for the XH component}
\description{
Implements the \strong{XH} component using a Susceptible-Infectious-Prophylaxis
(SIP) compartmental model of malaria infection dynamics.
Individuals in \eqn{P} are protected from infection but not from becoming
infectious; they re-enter susceptibility when prophylaxis wanes.
}
\section{State Variables}{

\describe{
\item{\code{H}}{total human (or host) population density}
\item{\code{I}}{density of infectious humans}
\item{\code{P}}{density of humans under chemoprophylaxis}
}
Note: susceptible density \eqn{S = H - I - P}.
}

\section{Parameters}{

\describe{
\item{\code{b}}{transmission probability from mosquito to human}
\item{\code{c}}{transmission probability from human to mosquito}
\item{\code{r}}{clearance rate for infections}
\item{\code{rho}}{fraction of new infections immediately treated / entering \eqn{P}}
\item{\code{xi}}{rate of mass drug administration (\eqn{\xi(t)})}
\item{\code{sigma}}{rate of mass screen-and-treat (\eqn{\sigma(t)})}
\item{\code{eta}}{rate of prophylaxis waning}
\item{\code{B}}{time-dependent birth rate function \eqn{B(t, H)}}
\item{\code{D}}{linear operator (matrix) for mortality, migration, aging, and transfers}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dH/dt &= B(t,H) + D \cdot H \\
dI/dt &= (1-\rho)hS - (r + \xi + \sigma)I + D \cdot I \\
dP/dt &= \rho hS + \xi(S+I) + \sigma I - \eta P + D \cdot P \\
\end{array}}
where \eqn{h} is the force of infection and \eqn{S = H - I - P}.
}

