% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{SIR}
\alias{SIR}
\title{The \code{SIR} module for the XH component}
\description{
Implements the \strong{XH} component using a Susceptible-Infectious-Recovered
(SIR) compartmental model of human infection dynamics.
}
\section{State Variables}{

\describe{
\item{\code{H}}{total human (or host) population density}
\item{\code{I}}{density of infectious humans}
\item{\code{R}}{density of recovered humans}
}
Note: susceptible density \eqn{S = H - I - R}.
}

\section{Parameters}{

\describe{
\item{\code{b}}{transmission probability from mosquito to human}
\item{\code{c}}{transmission probability from human to mosquito}
\item{\code{r}}{clearance rate for infections}
\item{\code{B}}{time-dependent birth rate function \eqn{B(t, H)}}
\item{\code{D}}{linear operator (matrix) for mortality, migration, aging, and transfers}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dH/dt &= B(t,H) + D \cdot H \\
dI/dt &= hS - rI + D \cdot I \\
dR/dt &= rI + D \cdot R \\
\end{array}}
where \eqn{h} is the force of infection and \eqn{S = H - I - R}.
}

