% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{skill_set_L.stages}
\alias{skill_set_L.stages}
\title{The \strong{L} module skill set}
\usage{
\method{skill_set_L}{stages}(Lname = "stages")
}
\arguments{
\item{Lname}{the name of the \strong{L} module}
}
\value{
\emph{L} module skill set, as a list
}
\description{
The \strong{L} skill set is a list of
a module's capabilities
}
\keyword{internal}
