% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{skill_set_MY.BQ}
\alias{skill_set_MY.BQ}
\title{The \strong{BQ} module skill set}
\usage{
\method{skill_set_MY}{BQ}(MYname)
}
\arguments{
\item{MYname}{the \strong{MY} module name}
}
\value{
\emph{MY} module skill set, as a list
}
\description{
The \strong{MY} skill set is a list of
a module's capabilities:
\itemize{
\item \code{demography} is
}
}
\keyword{internal}
