% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{skill_set_XH.SIPd}
\alias{skill_set_XH.SIPd}
\title{The \strong{XH} module skill set for \code{SIPd}}
\usage{
\method{skill_set_XH}{SIPd}(Xname = "SIP")
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
the skill set, as a list
}
\description{
The \strong{XH} skill set is a list of
a module's capabilities.
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
\keyword{internal}
