% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{skill_set_XH.SIR}
\alias{skill_set_XH.SIR}
\title{The \strong{XH} module skill set for \code{SIR}}
\usage{
\method{skill_set_XH}{SIR}(Xname = "SIR")
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
the skill set, as a list
}
\description{
The \strong{XH} skill set is a list of
a module's capabilities.
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
\keyword{internal}
