% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{skill_set_XH.SIRS}
\alias{skill_set_XH.SIRS}
\title{The \strong{XH} Module Skill Set}
\usage{
\method{skill_set_XH}{SIRS}(Xname = "SIRS")
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
the \code{SIRS} \emph{XH} module skill set, a list
}
\description{
The \strong{XH} skill set is a list of
an module's capabilities.
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
