% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{steady_state_L.stages}
\alias{steady_state_L.stages}
\title{Compute the steady state of \code{dLdt.stages} (\strong{L} component)}
\usage{
\method{steady_state_L}{stages}(eta, xds_obj, s = 1)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a named \link{list} with elements \code{E}, \code{L1}, \code{L2}, \code{L3}, \code{L4}, \code{P}
}
\description{
Given an egg deposition rate \code{eta},
return a steady state value for the equations in \link{dLdt.stages}.
Uses \link[stats:nlm]{stats::nlm} to minimise the sum of squared derivatives.
}
\note{
This function does not use deSolve
}
\keyword{internal}
