% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{steady_state_X.SEIR}
\alias{steady_state_X.SEIR}
\title{Compute the steady states for the SIRS model as a function of the daily EIR}
\usage{
\method{steady_state_X}{SEIR}(foi, H, xds_obj, i = 1)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the vector species index}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIRS model as a function of the daily eir.
}
\keyword{internal}
