% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{steady_state_X.SIP}
\alias{steady_state_X.SIP}
\title{Compute the steady states for the SIP model as a function of the daily foi}
\usage{
\method{steady_state_X}{SIP}(foi, H, xds_obj, i = 1)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the vector species index}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIP model as a function of the daily eir.
}
\keyword{internal}
