% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{steady_state_X.SIPd}
\alias{steady_state_X.SIPd}
\title{Compute the steady states for the  dts SIPd model as a function of the daily EIR}
\usage{
\method{steady_state_X}{SIPd}(foi, H, xds_obj, i = 1)

\method{steady_state_X}{SIPd}(foi, H, xds_obj, i = 1)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}
}
\value{
the steady states as a named vector

the steady states as a named vector
}
\description{
Compute the steady state of the  dts SIPd model as a function of the daily eir.

Compute the steady state of the SIPd model as a function of the daily eir.
}
