% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-weekly.R
\name{travel_malaria.weekly}
\alias{travel_malaria.weekly}
\title{A model for the travel Attack Rate}
\usage{
\method{travel_malaria}{weekly}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a [numeric]
}
\description{
Implements [travel_malaria] through a model for the travel Attack Rate
}
