% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{update_inits_X.SIPw}
\alias{update_inits_X.SIPw}
\title{Update inits for SIPw models from a vector of states}
\usage{
\method{update_inits_X}{SIPw}(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for SIPw models from a vector of states
}
