% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-dts-SIPw.R
\name{update_inits_X.SIPw_dts}
\alias{update_inits_X.SIPw_dts}
\title{Update inits for the SIPw_dts human model from a vector of states}
\usage{
\method{update_inits_X}{SIPw_dts}(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the SIPw_dts human model from a vector of states
}
