% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{update_inits_X.SIRnew}
\alias{update_inits_X.SIRnew}
\title{Update inits for the SIRnew human model from a vector of states}
\usage{
\method{update_inits_X}{SIRnew}(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the SIRnew human model from a vector of states
}
