% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{update_Linits.stages_dts}
\alias{update_Linits.stages_dts}
\title{Update inits for the stages_dts aquatic mosquito competition model}
\usage{
\method{update_Linits}{stages_dts}(pars, y, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the state variables}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for the stages_dts aquatic mosquito competition model
}
