% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{Update_MYt.SEI}
\alias{Update_MYt.SEI}
\title{Derivatives for adult mosquitoes}
\usage{
\method{Update_MYt}{SEI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:Update_MYt]{Update_MYt} for the SEI model.
}
\keyword{internal}
