% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-dts-ENBRQ.R
\name{update_MYZinits.ENBRQ_dts}
\alias{update_MYZinits.ENBRQ_dts}
\title{Make inits for ENBRQ_dts adult mosquito model}
\usage{
\method{update_MYZinits}{ENBRQ_dts}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a [list]
}
\description{
Make inits for ENBRQ_dts adult mosquito model
}
