% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{update_MYZinits.RMG}
\alias{update_MYZinits.RMG}
\title{Make inits for RMG adult mosquito model}
\usage{
\method{update_MYZinits}{RMG}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Make inits for RMG adult mosquito model
}
