% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-SBRQ.R
\name{update_MYZinits.SBRQ}
\alias{update_MYZinits.SBRQ}
\title{Make inits for SBRQ adult mosquito model}
\usage{
\method{update_MYZinits}{SBRQ}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Make inits for SBRQ adult mosquito model
}
