% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{update_Xinits.SIPmav3}
\alias{update_Xinits.SIPmav3}
\title{Update inits for SIPmav3 models from a vector of states}
\usage{
\method{update_Xinits}{SIPmav3}(pars, y, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the variables}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for SIPmav3 models from a vector of states
}
