% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{update_Xinits.SIRS}
\alias{update_Xinits.SIRS}
\title{Update inits for the SIRS human model from a vector of states}
\usage{
\method{update_Xinits}{SIRS}(pars, y, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the variables}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the SIRS human model from a vector of states
}
