% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-xde-workhorse.R
\name{update_Xinits.workhorse}
\alias{update_Xinits.workhorse}
\title{Update inits for the workhorse human model from a vector of states}
\usage{
\method{update_Xinits}{workhorse}(pars, y, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the variables}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the workhorse human model from a vector of states
}
