% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{Update_Xt.SEIRV}
\alias{Update_Xt.SEIRV}
\title{DTS updating for the SIS model for human / vertebrate host infections}
\usage{
\method{Update_Xt}{SEIRV}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [Update_Xt] for the SIS model
}
