% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{Update_Xt.SIPw}
\alias{Update_Xt.SIPw}
\title{Derivatives for human population}
\usage{
Update_Xt.SIPw(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [Update_Xt] for the SIPw dts model
}
