% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{Update_Xt.SIRS}
\alias{Update_Xt.SIRS}
\title{DTS updating for the SIS model for human / vertebrate host infections}
\usage{
\method{Update_Xt}{SIRS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Xt} for the SIS model
}
