% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{xde_lines_X_garki_xde}
\alias{xde_lines_X_garki_xde}
\title{Add lines for the density of infected individuals for the garki_xde model}
\usage{
xde_lines_X_garki_xde(XH, pars, clrs = viridisLite::turbo(7), llty = 1)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X_SIS}

\item{pars}{a list that defines an `ramp.xds` model (*e.g.*,  generated by `xde_setup()`)}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the garki_xde model
}
