% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{xde_lines_X_SEIR}
\alias{xde_lines_X_SEIR}
\title{Add lines for the density of infected individuals for the SEIR model}
\usage{
xde_lines_X_SEIR(
  XH,
  nStrata,
  clrs = c("black", "darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X_SEIR}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the SEIR model
}
