% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIRnew.R
\name{xde_lines_X_SEIRnew}
\alias{xde_lines_X_SEIRnew}
\title{Add lines for the density of infected individuals for the SEIRnew model}
\usage{
xde_lines_X_SEIRnew(
  XH,
  nStrata,
  clrs = c("black", "darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_deout_X_SEIRnew}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the SEIRnew model
}
