% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{xde_lines_X_SEIRV}
\alias{xde_lines_X_SEIRV}
\title{Add lines for the density of infected individuals for the SEIRV model}
\usage{
xde_lines_X_SEIRV(
  XH,
  nStrata,
  clrs = c("black", "darkblue", "darkred", "darkgreen", "purple"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X_SEIRV}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the SEIRV model
}
