% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIRnew.R
\name{xde_lines_X_SIRnew}
\alias{xde_lines_X_SIRnew}
\title{Add lines for the density of infected individuals for the SIRnew model}
\usage{
xde_lines_X_SIRnew(
  XH,
  nStrata,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_deout_X_SIRnew}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the SIRnew model
}
