% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{xde_setup_Xpar.garki_xde}
\alias{xde_setup_Xpar.garki_xde}
\title{xde_setup Xpar.garki_xde}
\usage{
\method{xde_setup_Xpar}{garki_xde}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [xde_setup_Xpar] for the garki_xde model
}
