% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{xde_steady_state_X.SEIRV}
\alias{xde_steady_state_X.SEIRV}
\title{Compute the steady states for the SEIRV model as a function of the daily EIR}
\usage{
\method{xde_steady_state_X}{SEIRV}(foi, H, Xpar)
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIS model as a function of the daily eir.
}
