% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{xde_steady_state_X.SIPmav3}
\alias{xde_steady_state_X.SIPmav3}
\title{Compute the steady states for the SIP model as a function of the daily foi}
\usage{
\method{xde_steady_state_X}{SIPmav3}(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{Xpar}{a list that defines an \eqn{\cal X} model}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIP model as a function of the daily eir.
}
