% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{xde_steady_state_X.SIRS}
\alias{xde_steady_state_X.SIRS}
\title{Compute the steady states for the SIRS model as a function of the daily EIR}
\usage{
\method{xde_steady_state_X}{SIRS}(foi, H, Xpar)

\method{xde_steady_state_X}{SIRS}(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{host density}

\item{Xpar}{a list that defines a X-class model}
}
\value{
the steady states as a named vector

the steady states as a named vector
}
\description{
Compute the steady state of the SIRS model as a function of the daily eir.

Compute the steady state of the SIS model as a function of the daily eir.
}
