% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-garki.R
\name{xds_lines_X_garki}
\alias{xds_lines_X_garki}
\title{Add lines for the density of infected individuals for the garki model}
\usage{
xds_lines_X_garki(times, XH, pars, clrs = viridisLite::turbo(8), llty = 1)
}
\arguments{
\item{times}{time points for the observations}

\item{XH}{a list with the outputs of parse_outputs_X_SIS}

\item{pars}{a list that defines an `ramp.xds` model (*e.g.*,  generated by `make()`)}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the garki model
}
