% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{xds_lines_X_SIR}
\alias{xds_lines_X_SIR}
\title{Add lines for the density of infected individuals for the SIR model}
\usage{
xds_lines_X_SIR(
  time,
  XH,
  nStrata,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{time}{time points for the observations}

\item{XH}{a list with the outputs of parse_outputs_X_SIR}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIR model
}
