% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{xds_lines_X_workhorse}
\alias{xds_lines_X_workhorse}
\title{Add lines for the density of infected individuals for the workhorse model}
\usage{
xds_lines_X_workhorse(
  time,
  XH,
  nStrata,
  clrs = c("darkblue", "darkgreen", "darkred", "purple"),
  llty = 1
)
}
\arguments{
\item{time}{a list with the outputs of parse_outputs_X_workhorse}

\item{XH}{a list with the outputs of parse_outputs_X_workhorse}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the workhorse model
}
