% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{xds_plot_X.SEIRV}
\alias{xds_plot_X.SEIRV}
\title{Plot the density of infected individuals for the SEIRV model}
\usage{
\method{xds_plot_X}{SEIRV}(
  xds_obj,
  i = 1,
  clrs = c("black", "darkblue", "darkred", "darkgreen", "purple"),
  llty = 1,
  add = FALSE
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Plot the density of infected individuals for the SEIRV model
}
