% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-SIPmav3.R
\name{xds_plot_X.SIPmav3}
\alias{xds_plot_X.SIPmav3}
\title{Plot the density of infected individuals for the SIPmav3 model}
\usage{
\method{xds_plot_X}{SIPmav3}(
  pars,
  i = 1,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1,
  add = FALSE
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Plot the density of infected individuals for the SIPmav3 model
}
