% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-xde-workhorse.R
\name{xds_plot_X.workhorse}
\alias{xds_plot_X.workhorse}
\title{Plot the density of infected individuals for the workhorse model}
\usage{
\method{xds_plot_X}{workhorse}(
  pars,
  i = 1,
  clrs = c("darkblue", "darkgreen", "darkred", "purple"),
  llty = 1,
  add = FALSE
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Plot the density of infected individuals for the workhorse model
}
