% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_dispersal.R
\name{change_K_matrix.Kq}
\alias{change_K_matrix.Kq}
\title{Change Mosquito Dispersal Matrix}
\usage{
\method{change_K_matrix}{Kq}(K_matrix, xds_obj, which_K = "Kq", s = 1)
}
\arguments{
\item{K_matrix}{a mosquito dispersal \link{matrix}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{which_K}{which K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run \link[ramp.xds:check_K_matrix]{ramp.xds::check_K_matrix} then

After passing checks, \code{xds_obj} is updated.

In models with multiple species, use \code{s} to
specify the species to update.
}
\seealso{
\link[ramp.xds:xds_info_mosquito_dispersal]{ramp.xds::xds_info_mosquito_dispersal}; \link[ramp.xds:setup_K_matrix]{ramp.xds::setup_K_matrix}
}
\keyword{internal}
