% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{change_XH_inits.garki}
\alias{change_XH_inits.garki}
\title{Set XH_inits.garki}
\usage{
\method{change_XH_inits}{garki}(xds_obj, i, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
a \link{list} vector
}
\description{
Implements \link{change_XH_inits} for the garki model
}
