% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{check_XH.SIRS}
\alias{check_XH.SIRS}
\title{Check / update before solving}
\usage{
\method{check_XH}{SIRS}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Check / update before solving
}
