% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{create_Xinits_SIPw}
\alias{create_Xinits_SIPw}
\title{Make initial values for a SIPw human model, with defaults}
\usage{
create_Xinits_SIPw(
  nStrata,
  Xopts = list(),
  H0 = NULL,
  S0 = NULL,
  I0 = 1,
  P0 = 0,
  w0 = 0
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{Xopts}{a [list] that could overwrite defaults}

\item{H0}{the initial human population density}

\item{S0}{the initial values of the variable S}

\item{I0}{the initial values of the variable I}

\item{P0}{the initial values of the variable P}

\item{w0}{the initial values of the tracking variable w}
}
\value{
a [list]
}
\description{
Make initial values for a SIPw human model, with defaults
}
