% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BRQS.R
\name{dMYdt.BRQS}
\alias{dMYdt.BRQS}
\title{Compute derivatives for \code{BRQS} (\strong{MY})}
\usage{
\method{dMYdt}{BRQS}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:dMYdt]{ramp.xds::dMYdt} for the BRQS ODE model.
}
\keyword{internal}
