% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMG.R
\name{dMYdt.RMG}
\alias{dMYdt.RMG}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYdt}{RMG}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYdt} for the RMG ODE model.
}
