% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{dMYdt.SEI}
\alias{dMYdt.SEI}
\title{\eqn{\cal MY} Component Derivatives for the \code{SEI} Mosquito Model}
\usage{
\method{dMYdt}{SEI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:dMYdt]{ramp.xds::dMYdt} for the SEI ODE model.
}
\details{
The dynamics of adult mosquitoes:
\deqn{\frac{dM}{dt} = \Lambda - \Omega \cdot M}

The density of infected but not infectious mosquitoes:
\deqn{\frac{dY}{dt} = fq\kappa(M-Y) - \Omega \cdot Y}

The density of infectious mosquitoes:
\deqn{\frac{dZ}{dt} = (Y-Z)/\tau - \Omega \cdot Z}
}
