% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{dXdt.SEIRV}
\alias{dXdt.SEIRV}
\title{Compute the derivatives for parasite infection dynamics in human population strata}
\usage{
\method{dXdt}{SEIRV}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SEIRV model
}
