% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEIS.R
\name{dXHdt.SEIS}
\alias{dXHdt.SEIS}
\title{\eqn{\cal X} Component Derivatives for the \code{SEIS} Model}
\usage{
\method{dXHdt}{SEIS}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXHdt} for the SEIS model
}
