% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{dXHdt.SIP}
\alias{dXHdt.SIP}
\title{Derivatives for the \code{SIP} Module (\strong{X} Component)}
\usage{
\method{dXHdt}{SIP}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for SIP compartmental model:
\deqn{
\begin{array}{rrrrcc}
dS/dt =& - (h +\xi) S &+ r I & + \eta P & + d{\cal H}(S) &+ B(t, H) \\
dI/dt =& (1-\rho) h S & - (r+\xi) I &&+ d{\cal H}(I) \\
dP/dt =& (\rho h +\xi)S & +\xi I & - \eta P &+ d{\cal H}(P)
\end{array}
}
where \eqn{H = S+I+P}; \eqn{B(t, H)} is the
time-dependent birth rate; and the \eqn{d{\cal H}}
operator computes derivatives for the demographic model \eqn{\cal H}.
The parameter \eqn{\xi} can be modified by setting \code{mass_treat} to a
non-zero function of time.
}
\seealso{
The parameters are defined in \link{make_XH_obj_SIP}
}
