% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_clearance.R
\name{F_clear}
\alias{F_clear}
\title{Natural Parasite Clearance}
\usage{
F_clear(y, Xpars)
}
\arguments{
\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute natural parasite clearance
in a population
}
