% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{F_fqM.BQ}
\alias{F_fqM.BQ}
\title{Blood feeding rate of the infective mosquito population}
\usage{
\method{F_fqM}{BQ}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link[ramp.xds:F_fqM]{ramp.xds::F_fqM} for the BQ model.
}
\keyword{internal}
