% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{F_H.garki_xde}
\alias{F_H.garki_xde}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{garki_xde}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_H] for the SIS model.
}
