% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{F_H.SEIR}
\alias{F_H.SEIR}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SEIR}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_H] for the SEIR model.
}
