% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{F_I.SEISd}
\alias{F_I.SEISd}
\title{Size of effective infectious human population}
\usage{
\method{F_I}{SEISd}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_I]{ramp.xds::F_I} for the SEISd model.
}
