% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{F_infectivity.SIP}
\alias{F_infectivity.SIP}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_infectivity}{SIP}(y, xds_obj, i)
}
\arguments{
\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_infectivity} for the SIP model.
}
