% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{F_ni.SEIRV}
\alias{F_ni.SEIRV}
\title{Compute the NI}
\usage{
\method{F_ni}{SEIRV}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ni} for the SEIS model.
}
