% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_ni.SEISd}
\alias{F_ni.SEISd}
\title{Compute the NI}
\usage{
\method{F_ni}{SEISd}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_ni] for the SEISd model.
}
