% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{F_ni.SIR}
\alias{F_ni.SIR}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_ni}{SIR}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ni} for the SIR model.
}
