% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{F_pfpr_by_pcr.SEISd}
\alias{F_pfpr_by_pcr.SEISd}
\title{Compute the prevalence of infection by pcr}
\usage{
\method{F_pfpr_by_pcr}{SEISd}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link[ramp.xds:F_prevalence]{ramp.xds::F_prevalence} for the SEISd model.
}
