% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{F_pfpr_by_rdt.SIPd}
\alias{F_pfpr_by_rdt.SIPd}
\title{Compute the prevalence of infection by RDT}
\usage{
<<<<<<< HEAD:man/F_pr_by_rdt.SIPd.Rd
F_pr_by_rdt.SIPd(vars, Xpar)
=======
\method{F_pfpr_by_rdt}{SIPd}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
>>>>>>> acc05432180b090912479adcca191d41600b0345:man/F_pfpr_by_rdt.SIPd.Rd
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the SIPd model.
}
