% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{F_pr.garki}
\alias{F_pr.garki}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{garki}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the true pr for the garki model.
}
