% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{F_pr.garki_xde}
\alias{F_pr.garki_xde}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{garki_xde}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Compute the true pr for the garki_xde model.
}
