% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-xde-SIPw.R
\name{F_pr.SIPw}
\alias{F_pr.SIPw}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SIPw(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for SIPw models
}
