% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{F_pr.SIPw}
\alias{F_pr.SIPw}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{SIPw}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_pr] for SIPw models
}
