% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{F_pr.SIRS}
\alias{F_pr.SIRS}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.SIRS(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIRS model.
}
