% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{F_pr.SIRS}
\alias{F_pr.SIRS}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{SIRS}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIRS model.
}
