% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-xde-workhorse.R
\name{F_pr.workhorse}
\alias{F_pr.workhorse}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
F_pr.workhorse(vars, Xpar)
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the workhorse model.
}
