% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{F_prevalence.garki}
\alias{F_prevalence.garki}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_prevalence}{garki}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the true pr for the garki model.
}
