% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{F_prevalence.SEIS}
\alias{F_prevalence.SEIS}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
<<<<<<< HEAD:man/F_pr.SEIS.Rd
F_pr.SEIS(vars, Xpar)
=======
\method{F_prevalence}{SEIS}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
>>>>>>> acc05432180b090912479adcca191d41600b0345:man/F_prevalence.SEIS.Rd
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the SEIS model.
}
