% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_prevalence.SEISd}
\alias{F_prevalence.SEISd}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
<<<<<<< HEAD:man/F_pr.SEISd.Rd
F_pr.SEISd(vars, Xpar)
=======
\method{F_prevalence}{SEISd}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
>>>>>>> acc05432180b090912479adcca191d41600b0345:man/F_prevalence.SEISd.Rd
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the SEISd model.
}
