% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_prevalence.SIP}
\alias{F_prevalence.SIP}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
<<<<<<< HEAD:man/F_pr.SIP.Rd
F_pr.SIP(vars, Xpar)
=======
\method{F_prevalence}{SIP}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
>>>>>>> acc05432180b090912479adcca191d41600b0345:man/F_prevalence.SIP.Rd
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the SIP model.
}
