% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{F_prevalence_pcr.SEIS}
\alias{F_prevalence_pcr.SEIS}
\title{Compute the prevalence of infection by pcr}
\usage{
<<<<<<< HEAD:man/F_pr_by_pcr.SEIS.Rd
F_pr_by_pcr.SEIS(vars, Xpar)
=======
F_prevalence_pcr.SEIS(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
>>>>>>> acc05432180b090912479adcca191d41600b0345:man/F_prevalence_pcr.SEIS.Rd
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the SEIS model.
}
