% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-dddn.R
\name{F_sigma.dddn}
\alias{F_sigma.dddn}
\title{Dawn, day, dusk, night model for the human fraction}
\usage{
\method{F_sigma}{dddn}(t, vars, sigma_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{sigma_par}{a \link{list}}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Implements [F_sigma] for a dddn model
}
