% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{get_HTC.garki}
\alias{get_HTC.garki}
\title{Compute the HTC for the garki model}
\usage{
\method{get_HTC}{garki}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{ramp.xds::get_HTC} for the garki model
}
\keyword{internal}
