% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{get_HTC.SEIRV}
\alias{get_HTC.SEIRV}
\title{Compute the HTC for the SEIRV model}
\usage{
\method{get_HTC}{SEIRV}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{ramp.xds::get_HTC} for the SEIRV model with demography.
}
\keyword{internal}
