% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{get_HTC.SIP}
\alias{get_HTC.SIP}
\title{Compute the HTC for the SIP model}
\usage{
\method{get_HTC}{SIP}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{ramp.xds::get_HTC} for the SIP model with demography.
}
\keyword{internal}
