% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{get_HTC.SIPd}
\alias{get_HTC.SIPd}
\title{Compute the HTC for the SIPd model}
\usage{
\method{get_HTC}{SIPd}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:get_HTC]{ramp.xds::get_HTC} for the SIPd model with demography.
}
\keyword{internal}
