% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-dts-stages.R
\name{get_Linits.stages_dts}
\alias{get_Linits.stages_dts}
\title{Return initial values as a vector}
\usage{
get_Linits.stages_dts(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_Linits} for the GeRM model.
}
