% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-dts-ENBRQ.R
\name{get_MYZinits.ENBRQ_dts}
\alias{get_MYZinits.ENBRQ_dts}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{ENBRQ_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_MYZinits} for the ENBRQ_dts model.
}
