% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZ-xde-RMG.R
\name{get_sigma.RMG}
\alias{get_sigma.RMG}
\title{Get the feeding rate}
\usage{
\method{get_sigma}{RMG}(pars, s = 1)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{s}{the vector species index}
}
\value{
y a [numeric] vector assigned the class "dynamic"
}
\description{
Get the feeding rate
}
