% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{get_Xpars.SEIS}
\alias{get_Xpars.SEIS}
\title{Return the parameters as a list}
\usage{
\method{get_Xpars}{SEIS}(pars, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Parameter values for the \eqn{i^{th}} host are
stored as \code{pars$Xpar[[i]]}. This returns the stored parameter
values as a list.
}
\seealso{
\link{make_Xpar_SEIS}
}
