% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-garki.R
\name{HTC.garki}
\alias{HTC.garki}
\title{Compute the HTC for the garki model}
\usage{
\method{HTC}{garki}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the garki model
}
