% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{HTC.SEIR}
\alias{HTC.SEIR}
\title{Compute the HTC for the SEIR model}
\usage{
\method{HTC}{SEIR}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SEIR model with demography.
}
