% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{HTC.SEIRV}
\alias{HTC.SEIRV}
\title{Compute the HTC for the SEIRV model}
\usage{
\method{HTC}{SEIRV}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a [numeric] vector
}
\description{
Implements [HTC] for the SEIRV model with demography.
}
