% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{HTC.SIR}
\alias{HTC.SIR}
\title{Compute the HTC for the SIR model}
\usage{
\method{HTC}{SIR}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIR model with demography.
}
