% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{HTC.SIRS}
\alias{HTC.SIRS}
\title{Compute the HTC for the SIRS model}
\usage{
\method{HTC}{SIRS}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIRS model with demography.
}
