% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{make_XH_obj_SIP}
\alias{make_XH_obj_SIP}
\title{Make parameters for SIP human model, with defaults}
\usage{
make_XH_obj_SIP(
  nStrata,
  options = list(),
  b = 0.55,
  r = 1/180,
  c = 0.15,
  rho = 0.1,
  eta = 1/25,
  xi = 1/365,
  sigma = 0.5,
  F_mass_treat = F_zero
)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{options}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{rho}{probability of successful treatment upon infection}

\item{eta}{prophylaxis waning rate}

\item{xi}{background treatment rate}

\item{sigma}{increased treatment rate while infected}

\item{F_mass_treat}{mass treatment rates as a function of time}
}
\value{
a \link{list}
}
\description{
Make parameters for SIP human model, with defaults
}
