% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{make_Xinits.SEIR}
\alias{make_Xinits.SEIR}
\title{Setup Xinits.SEIR}
\usage{
\method{make_Xinits}{SEIR}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xinits] for the SEIR model
}
