% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{make_Xinits.SIR}
\alias{make_Xinits.SIR}
\title{Setup Xinits.SIR}
\usage{
\method{make_Xinits}{SIR}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [make_Xinits] for the SIR model
}
