% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{make_Xinits_SIR}
\alias{make_Xinits_SIR}
\title{Make initial values for the SIR human model, with defaults}
\usage{
make_Xinits_SIR(nStrata, H, Xopts = list(), I = 1, R = 0)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial value for H}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{I}{the initial value for I}

\item{R}{the initial values for R}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIR human model, with defaults
}
