% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{make_Xinits_SIRS}
\alias{make_Xinits_SIRS}
\title{Make initial values for the SIRS human model, with defaults}
\usage{
make_Xinits_SIRS(nStrata, H, Xopts = list(), I = 1, R = 1)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial value for H}

\item{Xopts}{a [list] to overwrite defaults}

\item{I}{the initial value for I}

\item{R}{the initial values for R}
}
\value{
a [list]
}
\description{
Make initial values for the SIRS human model, with defaults
}
