% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{MBionomics.BQ}
\alias{MBionomics.BQ}
\title{Mosquito bionomics for \code{BQ} (\strong{MY})}
\usage{
\method{MBionomics}{BQ}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link[ramp.xds:MBionomics]{ramp.xds::MBionomics} for the BQ model
}
\keyword{internal}
