% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{MBionomics.SEI}
\alias{MBionomics.SEI}
\title{Mosquito bionomics for \code{SEI} (\strong{MY})}
\usage{
\method{MBionomics}{SEI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link[ramp.xds:MBionomics]{ramp.xds::MBionomics} for the SEI model
}
\keyword{internal}
