% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BRQS.R
\name{parse_MY_orbits.BRQS}
\alias{parse_MY_orbits.BRQS}
\title{Parse the output of deSolve and return variables for the BRQS model}
\usage{
\method{parse_MY_orbits}{BRQS}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:parse_MY_orbits]{ramp.xds::parse_MY_orbits} for the BRQS model
}
\keyword{internal}
