% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L-dts-stages.R
\name{parse_outputs_L.stages_dts}
\alias{parse_outputs_L.stages_dts}
\title{Parse the variable names for the stages_dts model}
\usage{
parse_outputs_L.stages_dts(outputs, pars, s)
}
\value{
[list]
}
\description{
Implements [parse_outputs_L] for stages_dts competition model.
}
