% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{parse_outputs_X.SIPw}
\alias{parse_outputs_X.SIPw}
\title{Parse the output of deSolve and return variables for SIPw models}
\usage{
\method{parse_outputs_X}{SIPw}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [parse_outputs_X] for SIPw models
}
