% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{parse_XH_orbits.SEISd}
\alias{parse_XH_orbits.SEISd}
\title{parse the output of deSolve and return variables for the SEISd model}
\usage{
\method{parse_XH_orbits}{SEISd}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve::deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:parse_XH_orbits]{parse_XH_orbits} for the SEISd model
}
\keyword{internal}
