% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{parse_XH_orbits.SIP}
\alias{parse_XH_orbits.SIP}
\title{Parse Solutions Matrix for the SIP model}
\usage{
\method{parse_XH_orbits}{SIP}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve::deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
SIP module variables as a named list
}
\description{
Parse Solutions Matrix for the SIP model

In \link[ramp.xds:xds_solve]{xds_solve}, after solving a system of equations,
the orbits for \eqn{H}, \eqn{I} and \eqn{P} are
pulled from the solutions matrix, and the model
also computes \eqn{S=H-I-P}. The variables are returned
as a named list.
}
\keyword{internal}
