% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{parse_XH_orbits.SIPd}
\alias{parse_XH_orbits.SIPd}
\title{Parse Solutions Matrix for the SIPd model}
\usage{
\method{parse_XH_orbits}{SIPd}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve::deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:parse_XH_orbits]{parse_XH_orbits} for the SIPd model
}
\keyword{internal}
