% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{parse_XH_orbits.SIRS}
\alias{parse_XH_orbits.SIRS}
\title{parse the output of deSolve and return variables for the SIRS model}
\usage{
\method{parse_XH_orbits}{SIRS}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve::deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link[ramp.xds:parse_XH_orbits]{ramp.xds::parse_XH_orbits} for the SIRS model
}
