% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIRV.R
\name{parse_Xorbits.SEIRV}
\alias{parse_Xorbits.SEIRV}
\title{Parse the output of deSolve and return variables for the SEIRV model}
\usage{
\method{parse_Xorbits}{SEIRV}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_Xorbits} for the SEIRV model
}
