% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIR.R
\name{parse_Xorbits.SIR}
\alias{parse_Xorbits.SIR}
\title{Parse the output of deSolve and return variables for the SIR model}
\usage{
\method{parse_Xorbits}{SIR}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements [parse_Xorbits] for the SIR model
}
