% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-xde-SEI.R
\name{set_MYZpars.SEI}
\alias{set_MYZpars.SEI}
\title{Return the parameters as a list}
\usage{
\method{set_MYZpars}{SEI}(pars, s = 1, MYZopts = list())

\method{set_MYZpars}{SEI}(pars, s = 1, MYZopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{MYZopts}{a named list}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$MYZpar[[s]]}.

This method dispatches on the type of \code{pars$MYZpar[[s]]}.
}
