% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-garki.R
\name{set_Xinits.garki}
\alias{set_Xinits.garki}
\title{Set Xinits.garki}
\usage{
\method{set_Xinits}{garki}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{Xopts}{a named list}
}
\value{
a \link{list} vector
}
\description{
Implements \link{set_Xinits} for the garki model
}
