% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-stages.R
\name{setup_L_obj.stages}
\alias{setup_L_obj.stages}
\title{Set up \code{stages} (\strong{L})}
\usage{
\method{setup_L_obj}{stages}(Lname, xds_obj, s, options = list())
}
\arguments{
\item{Lname}{the class name of the \strong{L} module}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a named list to configure \strong{\code{L_obj}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
The function sets up \code{L_obj} for the \eqn{s^{th}} species
by calling \link{make_L_obj_stages}
}
\seealso{
\link{make_L_obj_stages}
}
\keyword{internal}
