% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-BQ.R
\name{setup_MY_inits.BQ}
\alias{setup_MY_inits.BQ}
\title{Setup initial values for the BQ model}
\usage{
\method{setup_MY_inits}{BQ}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link[ramp.xds:setup_MY_inits]{ramp.xds::setup_MY_inits} for the RM model
}
\keyword{internal}
