% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{setup_MY_ix.SEI}
\alias{setup_MY_ix.SEI}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{SEI}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link[ramp.xds:setup_MY_ix]{ramp.xds::setup_MY_ix} for the SEI model.
}
