% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SEI.R
\name{setup_MY_obj.SEI}
\alias{setup_MY_obj.SEI}
\title{Setup MY_obj for the SEI model}
\usage{
\method{setup_MY_obj}{SEI}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{a \link{list}}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_obj} for the SEI model
}
