% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{setup_XH_inits.garki}
\alias{setup_XH_inits.garki}
\title{Setup XH_inits.garki}
\usage{
\method{setup_XH_inits}{garki}(xds_obj, H, i, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_XH_inits} for the garki model
}
