% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{setup_XH_inits.SEIRV}
\alias{setup_XH_inits.SEIRV}
\title{Setup XH_obj.SEIRV}
\usage{
\method{setup_XH_inits}{SEIRV}(xds_obj, H, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link[ramp.xds:setup_XH_inits]{ramp.xds::setup_XH_inits} for the SEIRV model
}
