% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{setup_XH_inits.SEISd}
\alias{setup_XH_inits.SEISd}
\title{Setup initial values for \code{SEISd}}
\usage{
\method{setup_XH_inits}{SEISd}(xds_obj, H, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link[ramp.xds:setup_XH_inits]{ramp.xds::setup_XH_inits} for the SEISd model
}
