% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIRS.R
\name{setup_XH_inits.SIRS}
\alias{setup_XH_inits.SIRS}
\title{Setup XH_obj.SIRS}
\usage{
\method{setup_XH_inits}{SIRS}(xds_obj, H, i, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_XH_inits} for the SIRS model
}
