% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIRV.R
\name{setup_XH_ix.SEIRV}
\alias{setup_XH_ix.SEIRV}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_XH_ix}{SEIRV}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{setup_XH_ix} for the SEIRV model.
}
