% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SEISd.R
\name{setup_XH_obj.SEISd}
\alias{setup_XH_obj.SEISd}
\title{Setup XH_obj.SEISd}
\usage{
\method{setup_XH_obj}{SEISd}(Xname, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_XH_obj} for the SEISd model
}
