% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{setup_XH_obj.SIP}
\alias{setup_XH_obj.SIP}
\title{Setup \code{XH_obj} for an \code{SIP}}
\usage{
\method{setup_XH_obj}{SIP}(Xname, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
a \link{list} vector
}
\description{
Set up the \code{SIP} model with parameters
}
