% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SIR.R
\name{setup_XH_obj.SIR}
\alias{setup_XH_obj.SIR}
\title{Setup XH_obj.SIR}
\usage{
\method{setup_XH_obj}{SIR}(Xname, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
a \link{list} vector
}
\description{
Implements \link[ramp.xds:setup_XH_obj]{ramp.xds::setup_XH_obj} for the SIR model
}
