% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-workhorse.R
\name{setup_Xinits.workhorse}
\alias{setup_Xinits.workhorse}
\title{Setup Xinits.workhorse}
\usage{
\method{setup_Xinits}{workhorse}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xinits] for the workhorse model
}
