% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-garki.R
\name{setup_Xpar.garki}
\alias{setup_Xpar.garki}
\title{make Xpar.garki}
\usage{
\method{setup_Xpar}{garki}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the garki model
}
