% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{setup_Xpar.SIP}
\alias{setup_Xpar.SIP}
\title{Setup `Xpar` for an `SIP`}
\usage{
\method{setup_Xpar}{SIP}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Set up the `SIP` model with parameters
}
