% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{setup_Xpar.SIPd}
\alias{setup_Xpar.SIPd}
\title{Setup `Xpar` for an `SIPd`}
\usage{
\method{setup_Xpar}{SIPd}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a [list] vector
}
\description{
Implements [setup_Xpar] for the SIPd model
}
