% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIPd.R
\name{SIPd}
\alias{SIPd}
\title{The \code{SIPd} module for the XH component}
\description{
Implements the \strong{XH} component using a delay Susceptible-Infectious-Prophylaxis
(SIPd) model of malaria infection dynamics. This is the delay differential
equation version of the SIP model, in which the duration of prophylaxis
\eqn{\eta} is modelled as a fixed delay.
}
\section{State Variables}{

\describe{
\item{\code{H}}{total human (or host) population density}
\item{\code{I}}{density of infectious humans}
\item{\code{P}}{density of humans under chemoprophylaxis}
}
Note: susceptible density \eqn{S = H - I - P}.
}

\section{Parameters}{

\describe{
\item{\code{b}}{transmission probability from mosquito to human}
\item{\code{c}}{transmission probability from human to mosquito}
\item{\code{r}}{clearance rate for infections}
\item{\code{rho}}{fraction of new infections immediately treated / entering \eqn{P}}
\item{\code{xi}}{rate of mass drug administration (\eqn{\xi})}
\item{\code{eta}}{prophylaxis duration (fixed delay, in days)}
\item{\code{B}}{time-dependent birth rate function \eqn{B(t, H)}}
\item{\code{D}}{linear operator (matrix) for mortality, migration, aging, and transfers}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dH/dt &= B(t,H) + D \cdot H \\
dI/dt &= (1-\rho)hS - (r + \xi)I + D \cdot I \\
dP/dt &= (\rho h + \xi)S + \xi I - \dot{\mathcal{H}}(P) + D \cdot P \\
\end{array}}
where \eqn{\dot{\mathcal{H}}(P)} is the delayed outflow from \eqn{P},
computed via \link[deSolve:timelags]{deSolve::lagderiv}.
}

