% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-xde-SEIR.R
\name{Update_XHt.SEIR}
\alias{Update_XHt.SEIR}
\title{DTS updating for the SIS model for human / vertebrate host infections}
\usage{
\method{Update_XHt}{SEIR}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:Update_XHt]{ramp.xds::Update_XHt} for the SIS model
}
