% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-XH-SIP.R
\name{Update_XHt.SIP}
\alias{Update_XHt.SIP}
\title{Derivatives for human population}
\usage{
\method{Update_XHt}{SIP}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link[ramp.xds:Update_XHt]{ramp.xds::Update_XHt} for the SIP model.
}
