% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-garki.R
\name{update_Xinits.garki}
\alias{update_Xinits.garki}
\title{Update inits for the garki model}
\usage{
\method{update_Xinits}{garki}(pars, y, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the variables}

\item{i}{the host species index}
}
\value{
an 'xds' object
}
\description{
Update inits for the garki model
}
