% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{update_Xinits.SIRS}
\alias{update_Xinits.SIRS}
\title{Update inits for the SIRS human model from a vector of states}
\usage{
\method{update_Xinits}{SIRS}(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the SIRS human model from a vector of states
}
