% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-SIPw.R
\name{xde_steady_state_X.SIPw}
\alias{xde_steady_state_X.SIPw}
\title{Compute the steady states for the SIP model as a function of the daily foi}
\usage{
\method{xde_steady_state_X}{SIPw}(foi, H, Xpar)
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIP model as a function of the daily eir.
}
