% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIRS.R
\name{xds_lines_X_SIRS}
\alias{xds_lines_X_SIRS}
\title{Add lines for the density of infected individuals for the SIRS model}
\usage{
xds_lines_X_SIRS(
  time,
  XH,
  nStrata,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{time}{time points for the observations}

\item{XH}{a list with the outputs of parse_outputs_X_SIRS}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the `lty` for plotting}
}
\description{
Add lines for the density of infected individuals for the SIRS model
}
