% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria-garki.R
\name{xds_plot_X.garki}
\alias{xds_plot_X.garki}
\title{Plot the density of infected individuals for the garki model}
\usage{
\method{xds_plot_X}{garki}(xds_obj, i = 1, clrs = viridisLite::turbo(7), llty = 1, add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Plot the density of infected individuals for the garki model
}
