% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SEIR.R
\name{xds_plot_X.SEIR}
\alias{xds_plot_X.SEIR}
\title{Plot the density of infected individuals for the SEIR model}
\usage{
\method{xds_plot_X}{SEIR}(
  pars,
  i = 1,
  clrs = c("black", "darkblue", "darkred", "darkgreen"),
  llty = 1,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the density of infected individuals for the SEIR model
}
